refactor(kernel): extract 16 kernel_handle trait impls into kernel::handles
Phase 1 of the kernel/mod.rs split. Move every `impl kernel_handle::* for
LibreFangKernel` block out of the 20K-line god-file into per-trait files
under `kernel/handles/`. The submodules are descendants of `kernel`, so
they retain access to LibreFangKernel's private fields and inherent
methods without any visibility surgery.

Mechanical, behavior-preserving move — no method bodies were touched.
mod.rs drops from ~20.6k to ~18.8k lines (-2073). Three imports that
became unused after the move were removed (kernel_handle self-alias,
ToolApprovalSubmission, std::str::FromStr) and re-imported locally in
the consuming handle files.

Traits moved:
  AgentControl, MemoryAccess, TaskQueue, EventBus, KnowledgeGraph,
  CronControl, HandsControl, ApprovalGate, A2ARegistry, ChannelSender,
  PromptStore, WorkflowRunner, GoalControl, ToolPolicy, ApiAuth,
  SessionWriter

Verified: cargo check -p librefang-kernel --lib + cargo clippy
-p librefang-kernel --lib -- -D warnings both green.
