+++ crates/librefang-kernel/src/kernel/mod.rs
@@ -23,7 +23,7 @@ use librefang_runtime::agent_loop::{
 };
 use librefang_runtime::audit::AuditLog;
 use librefang_runtime::drivers;
-use librefang_runtime::kernel_handle::{self, prelude::*};
+use librefang_runtime::kernel_handle::prelude::*;
 use librefang_runtime::llm_driver::{
     CompletionRequest, CompletionResponse, DriverConfig, LlmDriver, LlmError, StreamEvent,
 };
@@ -37,17 +37,23 @@ use librefang_types::config::{AuthProfile, AutoRouteStrategy, KernelConfig};
 use librefang_types::error::LibreFangError;
 use librefang_types::event::*;
 use librefang_types::memory::Memory;
-use librefang_types::tool::{AgentLoopSignal, ToolApprovalSubmission, ToolDefinition};
+use librefang_types::tool::{AgentLoopSignal, ToolDefinition};
 
 use arc_swap::ArcSwap;
 use async_trait::async_trait;
 use librefang_channels::types::SenderContext;
 use std::collections::HashSet;
 use std::path::{Path, PathBuf};
-use std::str::FromStr;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::{Arc, OnceLock, Weak};
 use tracing::{debug, error, info, instrument, warn};
+
+/// Per-trait `kernel_handle::*` impls live in their own files under
+/// `kernel/handles/` to keep this file from doubling as a trait-impl
+/// dumping ground. The submodules are descendants of `kernel`, so they
+/// retain access to `LibreFangKernel`'s private fields and inherent
+/// methods without any visibility surgery.
+mod handles;
 
 /// Synthetic `SenderContext.channel` value the cron dispatcher uses for
 /// `[[cron_jobs]]` fires. Matched in [`KernelHandle::resolve_user_tool_decision`]
@@ -133,194 +139,6 @@ pub(crate) fn resolve_cron_warn_threshold(
     let raw = (budget as f64) * frac;
     let threshold = raw.ceil() as u64;
     Some(threshold.min(budget))
-}
-
-/// Compute how many messages to keep in the cron session after applying both
-/// size caps, without mutating the slice.
-///
-/// Returns the number of messages (from the tail) that should survive. The
-/// caller uses this to determine the split point for either plain-drain or
-/// summarize-and-trim (#3693).
-pub(crate) fn cron_compute_keep_count(
-    messages: &[librefang_types::message::Message],
-    max_messages: Option<usize>,
-    max_tokens: Option<u64>,
-) -> usize {
-    use librefang_runtime::compactor::estimate_token_count;
-
-    let n = messages.len();
-    // Start with the full set; apply message-count cap first.
-    let after_msg_cap = if let Some(max_msgs) = max_messages {
-        n.min(max_msgs)
-    } else {
-        n
-    };
-
-    // Then apply token cap by trimming from the front.
-    if let Some(max_tok) = max_tokens {
-        // Linear scan from the cap downward (O(n) iterations × O(n) token estimation = O(n²);
-        // acceptable for cron session sizes seen in practice).
-        let mut keep = after_msg_cap;
-        while keep > 0 {
-            let start = n - keep;
-            let est = estimate_token_count(&messages[start..], None, None);
-            if est <= max_tok as usize {
-                break;
-            }
-            keep -= 1;
-        }
-        keep
-    } else {
-        after_msg_cap
-    }
-}
-
-/// Clamp the configured `cron_session_compaction_keep_recent` so that
-/// `[summary] + tail` never exceeds the size cap.
-///
-/// `keep_count` is the result of `cron_compute_keep_count` — i.e. how many
-/// messages the cap would allow on its own. After `SummarizeTrim`, the
-/// session contains `1 + tail_size` messages, so `tail_size` must be at
-/// most `keep_count - 1`. We also enforce a floor of `1` to keep at least
-/// one message of context after the summary.
-///
-/// Without this clamp, e.g. `cron_session_max_messages = 5` and
-/// `cron_session_compaction_keep_recent = 8` would yield `1 + 8 = 9`
-/// messages on the first fire (cap violated) and on subsequent fires
-/// `try_summarize_trim` would summarize a single message into a single
-/// summary, burning aux LLM calls without ever shrinking the session
-/// (#3693 review feedback on PR #4683).
-pub(crate) fn cron_clamp_keep_recent(keep_recent_cfg: usize, keep_count: usize) -> usize {
-    keep_recent_cfg.min(keep_count.saturating_sub(1)).max(1)
-}
-
-/// Decide which compaction mode to actually apply for this cron fire.
-///
-/// `SummarizeTrim` with `keep_count < 2` would write `[summary, last_msg]`
-/// (always 2 messages) into a session whose cap permits at most
-/// `keep_count` messages. The next fire would re-trigger `SummarizeTrim`
-/// and re-summarize 1 message into 1 summary forever, burning aux LLM
-/// calls without converging. We re-route those fires to `Prune` so the
-/// session shrinks deterministically.
-///
-/// `keep_count == 0` is the same shape — happens when even the single
-/// newest message exceeds `cron_session_max_tokens`. `Prune` empties the
-/// session in that case (which the LLM call would never do).
-pub(crate) fn cron_resolve_compaction_mode(
-    configured: librefang_types::config::CronCompactionMode,
-    keep_count: usize,
-) -> librefang_types::config::CronCompactionMode {
-    use librefang_types::config::CronCompactionMode;
-    match configured {
-        CronCompactionMode::SummarizeTrim if keep_count < 2 => CronCompactionMode::Prune,
-        m => m,
-    }
-}
-
-/// Apply a plain prune (drain-from-front) to a cron session, keeping the
-/// newest `session.messages.len() - drop_count` messages.
-///
-/// Used uniformly by the `Prune` main path, the `SummarizeTrim` fallback
-/// path, and the tool-pair-adjusted skip path so that all three paths
-/// produce identical side effects (`mark_messages_mutated`).
-fn apply_cron_prune(session: &mut librefang_memory::session::Session, drop_count: usize) {
-    if drop_count == 0 {
-        return;
-    }
-    session.messages.drain(0..drop_count);
-    session.mark_messages_mutated();
-}
-
-/// Attempt to summarize the messages that fall before `tail_start` using an
-/// LLM aux call, then stitch the result back as `[summary_msg] + tail`.
-///
-/// Returns `Some(new_messages)` on a real LLM success (non-empty, non-fallback
-/// summary). Returns `None` when the LLM fails or returns a fallback placeholder,
-/// so the caller can apply a plain prune instead.
-///
-/// The tool-pair boundary is adjusted with `adjust_split_for_tool_pair` before
-/// slicing so that `Assistant{ToolUse}` / `User{ToolResult}` pairs are never
-/// separated across the summary / tail boundary. If the adjustment pushes
-/// `tail_start` to `0` (nothing left to summarize) or to `messages.len()` (the
-/// keep_recent window consumed everything), `None` is returned immediately so the
-/// caller can decide whether to prune or skip.
-///
-/// Note: the caller holds the per-session mutex (`_prune_guard`) across this
-/// await, which spans the entire LLM summary call. This is intentional — it
-/// serialises `SummarizeTrim` runs on the same session so two concurrent fires
-/// cannot each start a summary against the same un-compacted snapshot.
-async fn try_summarize_trim(
-    messages: &[librefang_types::message::Message],
-    keep_recent: usize,
-    driver: std::sync::Arc<dyn librefang_runtime::llm_driver::LlmDriver>,
-    model: &str,
-) -> Option<Vec<librefang_types::message::Message>> {
-    use librefang_runtime::compactor::{
-        adjust_split_for_tool_pair, compact_messages, CompactionConfig,
-    };
-    use librefang_types::message::{Message, MessageContent, Role};
-
-    // Fast-fail when the caller could not resolve a model name (e.g. the agent
-    // disappeared from the registry between cron tick and prune). With an empty
-    // model name `compact_messages` would still walk stage-1 (with retries) →
-    // stage-2 (chunked) → stage-3 (fallback placeholder) and only then return
-    // `used_fallback = true`, holding the per-session mutex and one cron_lane
-    // slot for the full fail-out chain. Skip straight to the prune fallback.
-    if model.is_empty() {
-        return None;
-    }
-
-    let raw_tail_start = messages.len().saturating_sub(keep_recent);
-
-    // Adjust so we never split an Assistant{ToolUse} / User{ToolResult} pair.
-    let tail_start = adjust_split_for_tool_pair(messages, raw_tail_start, keep_recent);
-
-    // Nothing to summarize — skip gracefully.
-    if tail_start == 0 || tail_start == messages.len() {
-        return None;
-    }
-
-    let to_summarize = &messages[..tail_start];
-    let kept_tail = messages[tail_start..].to_vec();
-
-    // We've already split off the kept tail above; tell `compact_messages` to
-    // summarise the entire input it receives by setting `keep_recent = 0`.
-    //
-    // `max_retries = 1` (cron-only override): the per-session mutex and one
-    // cron_lane slot are held across this LLM call. The default of 3 retries
-    // inside stage-1 single-pass + the additional stage-2 chunked attempt can
-    // stretch to tens of seconds on a flaky provider, blocking sibling fires
-    // for the same `(agent, "cron")` session. Cron's failure mode is "fall
-    // back to plain prune", so a single attempt is sufficient — operators who
-    // want more aggressive retries should configure them in the aux driver
-    // itself, not amplify them here.
-    let compact_cfg = CompactionConfig {
-        threshold: 0,
-        keep_recent: 0,
-        max_retries: 1,
-        ..CompactionConfig::default()
-    };
-
-    // Only accept a real LLM summary — reject empty summaries and fallback
-    // placeholders (used_fallback = true means the LLM was unavailable and
-    // compact_messages substituted a synthetic "[Session compacted: ...]" stub).
-    match compact_messages(driver, model, to_summarize, &compact_cfg).await {
-        Ok(result) if !result.summary.is_empty() && !result.used_fallback => {
-            let summary_msg = Message {
-                role: Role::Assistant,
-                content: MessageContent::Text(format!(
-                    "[Cron session summary — {} messages compacted]\n\n{}",
-                    result.compacted_count, result.summary,
-                )),
-                pinned: false,
-                timestamp: None,
-            };
-            let mut new_messages = vec![summary_msg];
-            new_messages.extend(kept_tail);
-            Some(new_messages)
-        }
-        Ok(_) | Err(_) => None,
-    }
 }
 
 // ---------------------------------------------------------------------------
@@ -778,10 +596,6 @@ pub struct LibreFangKernel {
     pub(crate) scheduler: AgentScheduler,
     /// Memory substrate.
     pub(crate) memory: Arc<MemorySubstrate>,
-    /// Memory wiki vault (issue #3329). `None` when `[memory_wiki] enabled
-    /// = false`, in which case the `WikiAccess` trait methods short-circuit
-    /// to `KernelOpError::unavailable("wiki")`.
-    pub(crate) wiki_vault: Option<Arc<librefang_memory_wiki::WikiVault>>,
     /// Proactive memory store (mem0-style auto_retrieve/auto_memorize).
     pub(crate) proactive_memory: OnceLock<Arc<librefang_memory::ProactiveMemoryStore>>,
     /// Concrete handle to the LLM-backed memory extractor used by
@@ -2889,31 +2703,6 @@ impl LibreFangKernel {
             warn!("Config: {}", w);
         }
 
-        // Tool-exec backend (#3332): validate the active subtable is
-        // present and well-formed. A misconfigured `kind = "ssh"` /
-        // `kind = "daytona"` aborts boot here rather than failing on
-        // the first tool call.
-        if let Err(e) = config.tool_exec.validate() {
-            return Err(KernelError::BootFailed(format!(
-                "tool_exec config invalid: {e}"
-            )));
-        }
-        // Honesty banner: trait-route dispatch is staged behind the
-        // existing tool_runner.rs call sites. Operators who set
-        // `kind != "local"` should know they're configuring a backend
-        // that will become active in a follow-up PR.
-        if config.tool_exec.kind != librefang_types::tool_exec::BackendKind::Local {
-            warn!(
-                kind = %config.tool_exec.kind.as_str(),
-                "tool_exec.kind = '{}' configured, but tool dispatch via the new \
-                 ToolExecBackend trait is not yet wired into tool_runner.rs \
-                 (tracked as a follow-up). Tool calls will continue to use the \
-                 legacy local/docker path until that lands. \
-                 See docs/architecture/tool-exec-backends.md.",
-                config.tool_exec.kind.as_str()
-            );
-        }
-
         // Check TOTP configuration consistency
         if config.approval.second_factor == librefang_types::approval::SecondFactor::Totp {
             let vault_path = config.home_dir.join("vault.enc");
@@ -2953,11 +2742,10 @@ impl LibreFangKernel {
             .sqlite_path
             .clone()
             .unwrap_or_else(|| config.data_dir.join("librefang.db"));
-        let mut substrate = MemorySubstrate::open_with_pool_size(
+        let mut substrate = MemorySubstrate::open_with_chunking(
             &db_path,
             config.memory.decay_rate as f32,
             config.memory.chunking.clone(),
-            config.memory.pool_size,
         )
         .map_err(|e| KernelError::BootFailed(format!("Memory init failed: {e}")))?;
 
@@ -3335,14 +3123,13 @@ impl LibreFangKernel {
 
         // Initialize metering engine (shares the same SQLite connection as the memory substrate)
         let metering = Arc::new(MeteringEngine::new(Arc::new(
-            librefang_memory::usage::UsageStore::new(memory.pool()),
+            librefang_memory::usage::UsageStore::new(memory.usage_conn()),
         )));
 
         // Initialize prompt versioning and A/B experiment store with its own connection
         // to avoid conflicts with UsageStore concurrent writes
-        let prompt_store =
-            librefang_memory::PromptStore::new_with_path(&db_path, config.memory.pool_size)
-                .map_err(|e| KernelError::BootFailed(format!("Prompt store init failed: {e}")))?;
+        let prompt_store = librefang_memory::PromptStore::new_with_path(&db_path)
+            .map_err(|e| KernelError::BootFailed(format!("Prompt store init failed: {e}")))?;
 
         let supervisor = Supervisor::new();
         let background = BackgroundExecutor::with_concurrency(
@@ -3817,8 +3604,10 @@ impl LibreFangKernel {
         }
 
         // Initialize execution approval manager
-        let approval_manager =
-            crate::approval::ApprovalManager::new_with_db(config.approval.clone(), memory.pool());
+        let approval_manager = crate::approval::ApprovalManager::new_with_db(
+            config.approval.clone(),
+            memory.usage_conn(),
+        );
 
         // Validate notification config — warn (not error) on unrecognized values
         {
@@ -3908,28 +3697,6 @@ impl LibreFangKernel {
             None => config.data_dir.join("audit.anchor"),
         };
         let hooks_dir = config.home_dir.join("hooks");
-        // Optional memory-wiki vault (issue #3329). Off by default; only
-        // constructed when the operator has flipped `[memory_wiki] enabled
-        // = true`. A construction failure (e.g. unwritable vault path)
-        // logs a warning and disables the vault for this boot — it must
-        // not abort the kernel because the rest of the daemon is
-        // independent of the wiki feature.
-        let wiki_vault: Option<Arc<librefang_memory_wiki::WikiVault>> =
-            if config.memory_wiki.enabled {
-                match librefang_memory_wiki::WikiVault::new(&config.memory_wiki, &config.home_dir) {
-                    Ok(v) => Some(Arc::new(v)),
-                    Err(err) => {
-                        tracing::warn!(
-                            error = %err,
-                            "[memory_wiki] enabled but vault construction failed; \
-                             wiki tools will return KernelOpError::unavailable"
-                        );
-                        None
-                    }
-                }
-            } else {
-                None
-            };
         // Snapshot the initial taint rule registry into a shared
         // `Arc<ArcSwap<...>>`. This swap is the single source of truth read
         // by every connected MCP server's scanner — `Self::reload_config`
@@ -3980,7 +3747,6 @@ impl LibreFangKernel {
             session_stream_hub: Arc::new(crate::session_stream_hub::SessionStreamHub::new()),
             scheduler: AgentScheduler::new(),
             memory: memory.clone(),
-            wiki_vault: wiki_vault.clone(),
             proactive_memory: OnceLock::new(),
             proactive_memory_extractor: OnceLock::new(),
             prompt_store: OnceLock::new(),
@@ -3989,7 +3755,10 @@ impl LibreFangKernel {
             template_registry: WorkflowTemplateRegistry::new(),
             triggers: trigger_engine,
             background,
-            audit_log: Arc::new(AuditLog::with_db_anchored(memory.pool(), audit_anchor_path)),
+            audit_log: Arc::new(AuditLog::with_db_anchored(
+                memory.usage_conn(),
+                audit_anchor_path,
+            )),
             metering,
             // ArcSwap lets config_reload rebuild on `[llm.auxiliary]` edits
             // without invalidating any long-lived `Arc<Kernel>` handle.
@@ -4851,27 +4620,6 @@ system_prompt = "You are a helpful assistant."
         // sibling enforcement points (boot restore, hot reload,
         // update_manifest).
         validate_manifest_module_path(&manifest, &name)?;
-
-        // tool_exec backend (#3332): if the manifest pins a backend
-        // override, the matching subtable must exist on the global
-        // config — otherwise the agent would fail at the first tool
-        // call, not at spawn. We do this before locking the registry so
-        // a misconfigured override never advances past spawn.
-        if let Some(override_kind) = manifest.tool_exec_backend {
-            if let Err(e) = self
-                .config
-                .load()
-                .tool_exec
-                .validate_override(override_kind)
-            {
-                return Err(KernelError::LibreFang(LibreFangError::InvalidInput(
-                    format!(
-                        "agent {name:?} tool_exec_backend = {:?} but {e}",
-                        override_kind.as_str()
-                    ),
-                )));
-            }
-        }
 
         // Use a deterministic agent ID derived from the agent name so the
         // same agent gets the same UUID across daemon restarts. This preserves
@@ -12421,24 +12169,13 @@ system_prompt = "You are a helpful assistant."
         let old_cfg = self.config.load();
         use crate::config_reload::{should_apply_hot, validate_config_for_reload};
 
-        // Read and parse the on-disk config via the strict loader (#4664).
-        // Unlike `crate::config::load_config`, `try_load_config` returns `Err`
-        // on every failure mode (TOML syntax error, broken `include = [...]`
-        // chain, migration failure, deserialize-shape mismatch) instead of
-        // silently falling back to `KernelConfig::default()`. Without this,
-        // the diff-and-apply path below would treat the defaults as the
-        // operator's intent and wipe out their `default_model`,
-        // `provider_api_keys`, channels, etc. — the symptom the bug report
-        // describes (a duplicate `[web.searxng]` key in `config.toml` made
-        // the dashboard stop loading after the next 30 s reload tick).
-        //
-        // Surfacing `Err` here lets the watcher's
-        // `Err(e) => tracing::warn!("Config hot-reload failed: {e}")` branch
-        // fire and the `POST /api/config/reload` handler return 400, both
-        // without touching the live config.
+        // Read and parse config file (using load_config to process $include directives)
         let config_path = self.home_dir_boot.join("config.toml");
-        let mut new_config = crate::config::try_load_config(&config_path)
-            .map_err(|e| format!("Config reload failed; live config unchanged: {e}"))?;
+        let mut new_config = if config_path.exists() {
+            crate::config::load_config(Some(&config_path))
+        } else {
+            return Err("Config file not found".to_string());
+        };
 
         // Clamp bounds on the new config before validating or applying.
         // Initial boot calls clamp_bounds() at kernel construction time,
@@ -12447,17 +12184,9 @@ system_prompt = "You are a helpful assistant."
         // startup path normally corrects.
         new_config.clamp_bounds();
 
-        // Validate new config. Use the same `live config unchanged` wrapper as
-        // the strict-loader path so every reload-rejection error carries the
-        // operator-actionable pledge — both for log readability and so the
-        // integration helper `assert_reload_rejects_and_preserves_default_model`
-        // (api crate) can assert reload-boundary semantics with one substring
-        // regardless of which inner branch tripped.
+        // Validate new config
         if let Err(errors) = validate_config_for_reload(&new_config) {
-            return Err(format!(
-                "Config reload failed; live config unchanged: validation: {}",
-                errors.join("; ")
-            ));
+            return Err(format!("Validation failed: {}", errors.join("; ")));
         }
 
         // Build the reload plan against the live capability set so changes
@@ -14595,20 +14324,16 @@ system_prompt = "You are a helpful assistant."
                                         }
                                     };
 
-                                    // Prune (or summarize-and-trim) the persistent cron
-                                    // session before firing if the user has configured a
-                                    // size cap, and emit a tracing::warn! when the
-                                    // post-compaction size is approaching the provider
-                                    // context window (#3693).
+                                    // Prune the persistent cron session before firing
+                                    // if the user has configured a size cap, and emit
+                                    // a tracing::warn! when the post-prune size is
+                                    // approaching the provider context window (#3693).
                                     if !wants_new_session {
                                         let cfg_snap = kernel_job.config.load();
                                         let max_tokens_raw = cfg_snap.cron_session_max_tokens;
                                         let max_messages_raw = cfg_snap.cron_session_max_messages;
                                         let warn_fraction = cfg_snap.cron_session_warn_fraction;
                                         let warn_fallback = cfg_snap.cron_session_warn_total_tokens;
-                                        let compaction_mode = cfg_snap.cron_session_compaction_mode;
-                                        let keep_recent_cfg =
-                                            cfg_snap.cron_session_compaction_keep_recent.max(1);
                                         drop(cfg_snap);
                                         let max_messages =
                                             resolve_cron_max_messages(max_messages_raw);
@@ -14643,168 +14368,35 @@ system_prompt = "You are a helpful assistant."
                                                 kernel_job.memory.get_session(cron_sid)
                                             {
                                                 use librefang_runtime::compactor::estimate_token_count;
-                                                // Compute how many messages must be removed
-                                                // to satisfy both caps. We do this without
-                                                // mutating yet so SummarizeTrim can see the
-                                                // full "to-be-dropped" prefix.
-                                                let keep_count = cron_compute_keep_count(
-                                                    &session.messages,
-                                                    max_messages,
-                                                    max_tokens,
-                                                );
-                                                // "Did the cap demand a shrink?" — set once from
-                                                // the helper's verdict and never re-read inside the
-                                                // mutating arms. All three apply paths (SummarizeTrim
-                                                // success / fallback prune / configured prune) leave
-                                                // the session in a state that needs to be persisted
-                                                // when this is true.
-                                                let needs_compaction =
-                                                    keep_count < session.messages.len();
-
-                                                if needs_compaction {
-                                                    use librefang_types::config::CronCompactionMode;
-                                                    // #4683 review: re-route SummarizeTrim → Prune
-                                                    // when the cap is too tight for [summary] +
-                                                    // 1-msg-tail (keep_count < 2). Without this,
-                                                    // SummarizeTrim would always write 2 messages
-                                                    // back into a session whose cap permits at
-                                                    // most keep_count, and the next fire would
-                                                    // re-enter SummarizeTrim → endless aux LLM
-                                                    // round-trips with no convergence.
-                                                    let effective_compaction_mode =
-                                                        cron_resolve_compaction_mode(
-                                                            compaction_mode,
-                                                            keep_count,
-                                                        );
-                                                    if compaction_mode
-                                                        == CronCompactionMode::SummarizeTrim
-                                                        && effective_compaction_mode
-                                                            == CronCompactionMode::Prune
-                                                    {
-                                                        tracing::warn!(
-                                                            agent_id = %agent_id,
-                                                            session_id = %cron_sid,
-                                                            job = %job_name,
-                                                            keep_count,
-                                                            "cron SummarizeTrim: cap too tight for [summary] + tail (keep_count < 2); falling back to Prune"
-                                                        );
-                                                    }
-                                                    match effective_compaction_mode {
-                                                        CronCompactionMode::SummarizeTrim => {
-                                                            // Attempt LLM summarization of the
-                                                            // messages that would be dropped.
-                                                            // Falls back to plain prune on error
-                                                            // or when the LLM returns a fallback
-                                                            // placeholder (used_fallback=true).
-                                                            // adjust_split_for_tool_pair is applied
-                                                            // inside try_summarize_trim to avoid
-                                                            // cutting an Assistant{ToolUse} /
-                                                            // User{ToolResult} pair.
-                                                            let driver = kernel_job
-                                                                .aux_client
-                                                                .load()
-                                                                .driver_for(
-                                                                    librefang_types::config::AuxTask::Compression,
-                                                                );
-                                                            // Model: use the agent's model when
-                                                            // available, otherwise an empty string.
-                                                            // `try_summarize_trim` fast-fails on
-                                                            // empty model names so we skip the LLM
-                                                            // call (and the per-session mutex hold
-                                                            // it would imply) entirely and route
-                                                            // straight to the plain-prune fallback
-                                                            // below. Missing the agent from the
-                                                            // registry mid-cron is a symptom of a
-                                                            // registry / scheduler inconsistency
-                                                            // worth surfacing in logs.
-                                                            let model = match kernel_job
-                                                                .registry
-                                                                .get(agent_id)
-                                                            {
-                                                                Some(e) => {
-                                                                    librefang_runtime::agent_loop::strip_provider_prefix(
-                                                                        &e.manifest.model.model,
-                                                                        &e.manifest.model.provider,
-                                                                    )
-                                                                }
-                                                                None => {
-                                                                    tracing::warn!(
-                                                                        agent_id = %agent_id,
-                                                                        session_id = %cron_sid,
-                                                                        job = %job_name,
-                                                                        "cron SummarizeTrim: agent missing from registry; \
-                                                                         skipping LLM summary and falling back to plain prune"
-                                                                    );
-                                                                    String::new()
-                                                                }
-                                                            };
-
-                                                            // Clamp keep_recent so [summary] + tail
-                                                            // never exceeds the size cap (see
-                                                            // cron_clamp_keep_recent for rationale).
-                                                            let effective_keep_recent =
-                                                                cron_clamp_keep_recent(
-                                                                    keep_recent_cfg,
-                                                                    keep_count,
-                                                                );
-                                                            match try_summarize_trim(
-                                                                &session.messages,
-                                                                effective_keep_recent,
-                                                                driver,
-                                                                &model,
-                                                            )
-                                                            .await
-                                                            {
-                                                                Some(new_messages) => {
-                                                                    let kept = new_messages.len();
-                                                                    session
-                                                                        .set_messages(new_messages);
-                                                                    tracing::info!(
-                                                                        agent_id = %agent_id,
-                                                                        session_id = %cron_sid,
-                                                                        job = %job_name,
-                                                                        kept,
-                                                                        "cron session summarize-and-trim complete"
-                                                                    );
-                                                                }
-                                                                None => {
-                                                                    // LLM unavailable, returned a
-                                                                    // fallback placeholder, or the
-                                                                    // tool-pair adjustment left
-                                                                    // nothing to summarize —
-                                                                    // fall back to plain prune so
-                                                                    // the fire is not blocked.
-                                                                    tracing::warn!(
-                                                                        agent_id = %agent_id,
-                                                                        session_id = %cron_sid,
-                                                                        job = %job_name,
-                                                                        "cron SummarizeTrim: LLM summarization failed or returned empty; \
-                                                                         falling back to Prune"
-                                                                    );
-                                                                    let drop_count =
-                                                                        session.messages.len()
-                                                                            - keep_count;
-                                                                    apply_cron_prune(
-                                                                        &mut session,
-                                                                        drop_count,
-                                                                    );
-                                                                }
-                                                            }
-                                                        }
-                                                        CronCompactionMode::Prune => {
-                                                            // Plain drop-from-front.
-                                                            let drop_count =
-                                                                session.messages.len() - keep_count;
-                                                            apply_cron_prune(
-                                                                &mut session,
-                                                                drop_count,
-                                                            );
-                                                        }
+                                                let mut mutated = false;
+                                                if let Some(max_msgs) = max_messages {
+                                                    if session.messages.len() > max_msgs {
+                                                        let excess =
+                                                            session.messages.len() - max_msgs;
+                                                        session.messages.drain(0..excess);
+                                                        session.mark_messages_mutated();
+                                                        mutated = true;
                                                     }
                                                 }
-
-                                                // Post-compaction approach-warn (#3693):
-                                                // estimate once after any changes so
+                                                if let Some(max_tok) = max_tokens {
+                                                    loop {
+                                                        let est = estimate_token_count(
+                                                            &session.messages,
+                                                            None,
+                                                            None,
+                                                        );
+                                                        if est <= max_tok as usize
+                                                            || session.messages.is_empty()
+                                                        {
+                                                            break;
+                                                        }
+                                                        session.messages.remove(0);
+                                                        session.mark_messages_mutated();
+                                                        mutated = true;
+                                                    }
+                                                }
+                                                // Post-prune approach-warn (#3693):
+                                                // estimate once after any drains so
                                                 // operators see the trend before the
                                                 // provider returns 400. Estimate
                                                 // omits system_prompt / tools — those
@@ -14820,18 +14412,6 @@ system_prompt = "You are a helpful assistant."
                                                         as u64;
                                                     if estimated >= threshold {
                                                         let budget = max_tokens.or(warn_fallback);
-                                                        // `post_compaction` distinguishes "we just
-                                                        // shrank and the session is still over the
-                                                        // soft threshold" (real signal — the
-                                                        // current fire's content is large) from
-                                                        // "the session was already over threshold
-                                                        // before any compaction" (operator should
-                                                        // tighten the cap). After SummarizeTrim
-                                                        // succeeds, the synthetic summary message
-                                                        // can itself be large enough to keep the
-                                                        // estimate above threshold, so this warn
-                                                        // landing right after a successful
-                                                        // compaction is expected — not a bug.
                                                         tracing::warn!(
                                                             agent_id = %agent_id,
                                                             session_id = %cron_sid,
@@ -14840,7 +14420,6 @@ system_prompt = "You are a helpful assistant."
                                                             threshold = threshold,
                                                             budget = ?budget,
                                                             messages = session.messages.len(),
-                                                            post_compaction = needs_compaction,
                                                             "cron session approaching context budget — \
                                                              consider lowering cron_session_max_tokens, \
                                                              enabling cron_session_max_messages, or \
@@ -14848,7 +14427,7 @@ system_prompt = "You are a helpful assistant."
                                                         );
                                                     }
                                                 }
-                                                if needs_compaction {
+                                                if mutated {
                                                     let _ = kernel_job
                                                         .memory
                                                         .save_session_async(&session)
@@ -18512,2192 +18091,29 @@ impl LibreFangKernel {
 }
 
 // ---- BEGIN role-trait impls (split from former `impl KernelHandle for LibreFangKernel`, #3746) ----
-
-#[async_trait::async_trait]
-impl kernel_handle::AgentControl for LibreFangKernel {
-    async fn spawn_agent(
-        &self,
-        manifest_toml: &str,
-        parent_id: Option<&str>,
-    ) -> Result<(String, String), kernel_handle::KernelOpError> {
-        // Verify manifest integrity if a signed manifest hash is present
-        let content_hash = librefang_types::manifest_signing::hash_manifest(manifest_toml);
-        tracing::debug!(hash = %content_hash, "Manifest SHA-256 computed for integrity tracking");
-
-        let manifest: AgentManifest =
-            toml::from_str(manifest_toml).map_err(|e| format!("Invalid manifest: {e}"))?;
-        let name = manifest.name.clone();
-        let parent = parent_id.and_then(|pid| pid.parse::<AgentId>().ok());
-        let id = self
-            .spawn_agent_with_parent(manifest, parent)
-            .map_err(|e| format!("Spawn failed: {e}"))?;
-        Ok((id.to_string(), name))
-    }
-
-    async fn send_to_agent(
-        &self,
-        agent_id: &str,
-        message: &str,
-    ) -> Result<String, kernel_handle::KernelOpError> {
-        let id = self.resolve_agent_identifier(agent_id)?;
-        let result = self
-            .send_message(id, message)
-            .await
-            .map_err(|e| format!("Send failed: {e}"))?;
-        Ok(result.response)
-    }
-
-    async fn send_to_agent_as(
-        &self,
-        agent_id: &str,
-        message: &str,
-        parent_agent_id: &str,
-    ) -> Result<String, kernel_handle::KernelOpError> {
-        let id = self.resolve_agent_identifier(agent_id)?;
-        // Parent resolution: try the name/alias resolver first for ergonomics,
-        // but fall back to bare UUID parsing when the parent has been removed
-        // from the registry. A parent can legitimately disappear from the
-        // registry mid-flight (e.g. /kill racing with a pending agent_send
-        // response), while its `SessionInterrupt` is still live in
-        // `session_interrupts` because the in-flight turn holds a clone.
-        // Failing here would break the cascade contract "parent absent →
-        // no cascade but call proceeds" that `send_message_as` implements.
-        let parent_id = self
-            .resolve_agent_identifier(parent_agent_id)
-            .or_else(|_| {
-                parent_agent_id
-                    .parse::<AgentId>()
-                    .map_err(|e| format!("bad parent_agent_id: {e}"))
-            })?;
-        let result = self
-            .send_message_as(id, message, parent_id)
-            .await
-            .map_err(|e| format!("Send failed: {e}"))?;
-        Ok(result.response)
-    }
-
-    fn list_agents(&self) -> Vec<kernel_handle::AgentInfo> {
-        self.registry
-            .list()
-            .into_iter()
-            .map(|e| kernel_handle::AgentInfo {
-                id: e.id.to_string(),
-                name: e.name.clone(),
-                state: format!("{:?}", e.state),
-                model_provider: e.manifest.model.provider.clone(),
-                model_name: e.manifest.model.model.clone(),
-                description: e.manifest.description.clone(),
-                tags: e.tags.clone(),
-                tools: e.manifest.capabilities.tools.clone(),
-            })
-            .collect()
-    }
-
-    fn touch_heartbeat(&self, agent_id: &str) {
-        if let Ok(id) = agent_id.parse::<AgentId>() {
-            self.registry.touch(id);
-        }
-    }
-
-    async fn run_forked_agent_oneshot(
-        &self,
-        agent_id: &str,
-        prompt: &str,
-        allowed_tools: Option<Vec<String>>,
-    ) -> Result<String, kernel_handle::KernelOpError> {
-        let id = agent_id
-            .parse::<AgentId>()
-            .map_err(|e| format!("bad agent_id: {e}"))?;
-        // Need `Arc<Self>` to call `run_forked_agent_streaming` (the method
-        // is defined on `Arc<LibreFangKernel>`). Upgrade via `self_handle`;
-        // if the weak ref is stale the daemon is shutting down and the
-        // extractor should abort.
-        let kernel = self
-            .self_handle
-            .get()
-            .and_then(|w| w.upgrade())
-            .ok_or_else(|| "kernel Arc unavailable (shutting down?)".to_string())?;
-        let (mut rx, handle) = kernel
-            .run_forked_agent_streaming(id, prompt, allowed_tools)
-            .map_err(|e| format!("fork start failed: {e}"))?;
-        // Drain the stream — we don't need streaming semantics for a
-        // one-shot completion, just the final text. The spawned task
-        // keeps running until `ContentComplete` (or error/abort) anyway.
-        while (rx.recv().await).is_some() {
-            // Events consumed; the final text is on the join handle's
-            // `AgentLoopResult.response`. Discarding these events is
-            // fine because `ContentComplete` is already signalled to
-            // the join handle by the time we observe channel close.
-        }
-        let result = handle
-            .await
-            .map_err(|e| format!("fork join failed: {e}"))?
-            .map_err(|e| format!("fork loop failed: {e}"))?;
-        Ok(result.response)
-    }
-
-    fn kill_agent(&self, agent_id: &str) -> Result<(), kernel_handle::KernelOpError> {
-        let id = self
-            .resolve_agent_identifier(agent_id)
-            .map_err(kernel_handle::KernelOpError::Internal)?;
-        LibreFangKernel::kill_agent(self, id)
-            .map_err(|e| kernel_handle::KernelOpError::Internal(format!("Kill failed: {e}")))
-    }
-
-    fn find_agents(&self, query: &str) -> Vec<kernel_handle::AgentInfo> {
-        let q = query.to_lowercase();
-        self.registry
-            .list()
-            .into_iter()
-            .filter(|e| {
-                let name_match = e.name.to_lowercase().contains(&q);
-                let tag_match = e.tags.iter().any(|t| t.to_lowercase().contains(&q));
-                let tool_match = e
-                    .manifest
-                    .capabilities
-                    .tools
-                    .iter()
-                    .any(|t| t.to_lowercase().contains(&q));
-                let desc_match = e.manifest.description.to_lowercase().contains(&q);
-                name_match || tag_match || tool_match || desc_match
-            })
-            .map(|e| kernel_handle::AgentInfo {
-                id: e.id.to_string(),
-                name: e.name.clone(),
-                state: format!("{:?}", e.state),
-                model_provider: e.manifest.model.provider.clone(),
-                model_name: e.manifest.model.model.clone(),
-                description: e.manifest.description.clone(),
-                tags: e.tags.clone(),
-                tools: e.manifest.capabilities.tools.clone(),
-            })
-            .collect()
-    }
-
-    async fn spawn_agent_checked(
-        &self,
-        manifest_toml: &str,
-        parent_id: Option<&str>,
-        parent_caps: &[librefang_types::capability::Capability],
-    ) -> Result<(String, String), kernel_handle::KernelOpError> {
-        // Parse the child manifest to extract its capabilities
-        let child_manifest: AgentManifest = toml::from_str(manifest_toml)
-            .map_err(|e| kernel_handle::KernelOpError::InvalidInput(format!("manifest: {e}")))?;
-        let child_caps = manifest_to_capabilities(&child_manifest);
-
-        // Enforce: child capabilities must be a subset of parent capabilities
-        librefang_types::capability::validate_capability_inheritance(parent_caps, &child_caps)
-            .map_err(kernel_handle::KernelOpError::Internal)?;
-
-        tracing::info!(
-            parent = parent_id.unwrap_or("kernel"),
-            child = %child_manifest.name,
-            child_caps = child_caps.len(),
-            "Capability inheritance validated — spawning child agent"
-        );
-
-        // Delegate to the normal spawn path via the AgentControl role trait.
-        kernel_handle::AgentControl::spawn_agent(self, manifest_toml, parent_id).await
-    }
-
-    fn max_agent_call_depth(&self) -> u32 {
-        let cfg = self.config.load();
-        cfg.max_agent_call_depth
-    }
-
-    fn fire_agent_step(&self, agent_id: &str, step: u32) {
-        self.external_hooks.fire(
-            crate::hooks::ExternalHookEvent::AgentStep,
-            serde_json::json!({
-                "agent_id": agent_id.to_string(),
-                "step": step,
-            }),
-        );
-    }
-}
-
-impl kernel_handle::MemoryAccess for LibreFangKernel {
-    fn memory_store(
-        &self,
-        key: &str,
-        value: serde_json::Value,
-        peer_id: Option<&str>,
-    ) -> Result<(), kernel_handle::KernelOpError> {
-        use kernel_handle::KernelOpError;
-        let agent_id = shared_memory_agent_id();
-        let scoped = peer_scoped_key(key, peer_id);
-        // Check whether key already exists to determine Created vs Updated
-        let had_old = self
-            .memory
-            .structured_get(agent_id, &scoped)
-            .ok()
-            .flatten()
-            .is_some();
-        self.memory
-            .structured_set(agent_id, &scoped, value)
-            .map_err(|e| KernelOpError::Internal(format!("Memory store failed: {e}")))?;
-
-        // Publish MemoryUpdate event so triggers can react
-        let operation = if had_old {
-            MemoryOperation::Updated
-        } else {
-            MemoryOperation::Created
-        };
-        let event = Event::new(
-            agent_id,
-            EventTarget::Broadcast,
-            EventPayload::MemoryUpdate(MemoryDelta {
-                operation,
-                key: scoped.clone(),
-                agent_id,
-            }),
-        );
-        if let Some(weak) = self.self_handle.get() {
-            if let Some(kernel) = weak.upgrade() {
-                // Propagate trigger-chain depth across the spawn boundary
-                // (#3735). Without this, a memory_store invoked from inside
-                // a triggered agent would publish into a fresh top-level
-                // depth=0 scope, defeating the depth cap on chains that
-                // travel through memory updates.
-                let parent_depth = PUBLISH_EVENT_DEPTH.try_with(|c| c.get()).unwrap_or(0);
-                spawn_logged(
-                    "memory_event_publish",
-                    PUBLISH_EVENT_DEPTH.scope(std::cell::Cell::new(parent_depth), async move {
-                        kernel.publish_event(event).await;
-                    }),
-                );
-            }
-        }
-        Ok(())
-    }
-
-    fn memory_recall(
-        &self,
-        key: &str,
-        peer_id: Option<&str>,
-    ) -> Result<Option<serde_json::Value>, kernel_handle::KernelOpError> {
-        use kernel_handle::KernelOpError;
-        let agent_id = shared_memory_agent_id();
-        let scoped = peer_scoped_key(key, peer_id);
-        self.memory
-            .structured_get(agent_id, &scoped)
-            .map_err(|e| KernelOpError::Internal(format!("Memory recall failed: {e}")))
-    }
-
-    fn memory_list(
-        &self,
-        peer_id: Option<&str>,
-    ) -> Result<Vec<String>, kernel_handle::KernelOpError> {
-        use kernel_handle::KernelOpError;
-        let agent_id = shared_memory_agent_id();
-        let all_keys = self
-            .memory
-            .list_keys(agent_id)
-            .map_err(|e| KernelOpError::Internal(format!("Memory list failed: {e}")))?;
-        match peer_id {
-            Some(pid) => {
-                let prefix = format!("peer:{pid}:");
-                Ok(all_keys
-                    .into_iter()
-                    .filter_map(|k| k.strip_prefix(&prefix).map(|s| s.to_string()))
-                    .collect())
-            }
-            None => {
-                // When no peer context, return only non-peer-scoped keys
-                Ok(all_keys
-                    .into_iter()
-                    .filter(|k| !k.starts_with("peer:"))
-                    .collect())
-            }
-        }
-    }
-
-    fn memory_acl_for_sender(
-        &self,
-        sender_id: Option<&str>,
-        channel: Option<&str>,
-    ) -> Option<librefang_types::user_policy::UserMemoryAccess> {
-        if !self.auth.is_enabled() {
-            return None;
-        }
-        let user_id = self.auth.resolve_user(sender_id, channel)?;
-        self.auth.memory_acl_for(user_id)
-    }
-}
-
-impl kernel_handle::WikiAccess for LibreFangKernel {
-    fn wiki_get(&self, topic: &str) -> Result<serde_json::Value, kernel_handle::KernelOpError> {
-        use kernel_handle::KernelOpError;
-        let vault = self
-            .wiki_vault
-            .as_ref()
-            .ok_or_else(|| KernelOpError::unavailable("wiki_get"))?;
-        match vault.get(topic) {
-            Ok(page) => serde_json::to_value(&page)
-                .map_err(|e| KernelOpError::Internal(format!("Wiki get serialize: {e}"))),
-            Err(librefang_memory_wiki::WikiError::NotFound(_)) => Err(KernelOpError::Internal(
-                format!("wiki topic `{topic}` not found"),
-            )),
-            Err(err) => Err(KernelOpError::Internal(format!("Wiki get failed: {err}"))),
-        }
-    }
-
-    fn wiki_search(
-        &self,
-        query: &str,
-        limit: usize,
-    ) -> Result<serde_json::Value, kernel_handle::KernelOpError> {
-        use kernel_handle::KernelOpError;
-        let vault = self
-            .wiki_vault
-            .as_ref()
-            .ok_or_else(|| KernelOpError::unavailable("wiki_search"))?;
-        let hits = vault
-            .search(query, limit)
-            .map_err(|e| KernelOpError::Internal(format!("Wiki search failed: {e}")))?;
-        serde_json::to_value(&hits)
-            .map_err(|e| KernelOpError::Internal(format!("Wiki search serialize: {e}")))
-    }
-
-    fn wiki_write(
-        &self,
-        topic: &str,
-        body: &str,
-        provenance: serde_json::Value,
-        force: bool,
-    ) -> Result<serde_json::Value, kernel_handle::KernelOpError> {
-        use kernel_handle::KernelOpError;
-        let vault = self
-            .wiki_vault
-            .as_ref()
-            .ok_or_else(|| KernelOpError::unavailable("wiki_write"))?;
-        let prov: librefang_memory_wiki::ProvenanceEntry = serde_json::from_value(provenance)
-            .map_err(|e| {
-                KernelOpError::InvalidInput(format!(
-                    "wiki_write `provenance` must be {{agent, [session], [channel], [turn], at}}: {e}"
-                ))
-            })?;
-        match vault.write(topic, body, prov, force) {
-            Ok(outcome) => serde_json::to_value(&outcome)
-                .map_err(|e| KernelOpError::Internal(format!("Wiki write serialize: {e}"))),
-            Err(librefang_memory_wiki::WikiError::HandEditConflict { topic }) => {
-                Err(KernelOpError::Internal(format!(
-                    "wiki page `{topic}` was edited externally; re-read the file or pass force=true"
-                )))
-            }
-            Err(librefang_memory_wiki::WikiError::InvalidTopic { topic, reason }) => Err(
-                KernelOpError::InvalidInput(format!("wiki_write topic `{topic}`: {reason}")),
-            ),
-            Err(librefang_memory_wiki::WikiError::BodyTooLarge { topic, size, cap }) => {
-                Err(KernelOpError::InvalidInput(format!(
-                    "wiki_write body for `{topic}` is {size} bytes; exceeds the {cap}-byte cap"
-                )))
-            }
-            Err(err) => Err(KernelOpError::Internal(format!("Wiki write failed: {err}"))),
-        }
-    }
-}
-
-#[async_trait::async_trait]
-impl kernel_handle::TaskQueue for LibreFangKernel {
-    async fn task_post(
-        &self,
-        title: &str,
-        description: &str,
-        assigned_to: Option<&str>,
-        created_by: Option<&str>,
-    ) -> Result<String, kernel_handle::KernelOpError> {
-        use kernel_handle::KernelOpError;
-        let task_id = self
-            .memory
-            .task_post(title, description, assigned_to, created_by)
-            .await
-            .map_err(|e| KernelOpError::Internal(format!("Task post failed: {e}")))?;
-
-        let event = librefang_types::event::Event::new(
-            AgentId::new(), // system-originated
-            librefang_types::event::EventTarget::Broadcast,
-            librefang_types::event::EventPayload::System(
-                librefang_types::event::SystemEvent::TaskPosted {
-                    task_id: task_id.clone(),
-                    title: title.to_string(),
-                    assigned_to: assigned_to.map(String::from),
-                    created_by: created_by.map(String::from),
-                },
-            ),
-        );
-        self.publish_event(event).await;
-
-        Ok(task_id)
-    }
-
-    async fn task_claim(
-        &self,
-        agent_id: &str,
-    ) -> Result<Option<serde_json::Value>, kernel_handle::KernelOpError> {
-        use kernel_handle::KernelOpError;
-        // Resolve `agent_id` to a canonical UUID and also capture the name.
-        // Both are forwarded to `memory.task_claim` so that tasks whose
-        // `assigned_to` field was stored as either a UUID *or* a name string
-        // are correctly matched (issue #2841).
-        let (resolved, resolved_name) = match librefang_types::agent::AgentId::from_str(agent_id) {
-            Ok(parsed_id) => {
-                // Caller passed a UUID — look up the name from the registry.
-                let name = self.registry.get(parsed_id).map(|e| e.name.clone());
-                (agent_id.to_string(), name)
-            }
-            Err(_) => match self.registry.find_by_name(agent_id) {
-                Some(entry) => (entry.id.to_string(), Some(agent_id.to_string())),
-                None => {
-                    return Err(KernelOpError::AgentNotFound(agent_id.to_string()));
-                }
-            },
-        };
-        let result = self
-            .memory
-            .task_claim(&resolved, resolved_name.as_deref())
-            .await
-            .map_err(|e| KernelOpError::Internal(format!("Task claim failed: {e}")))?;
-
-        if let Some(ref task) = result {
-            let task_id = task["id"].as_str().unwrap_or("").to_string();
-            let event = librefang_types::event::Event::new(
-                AgentId::new(), // system-originated
-                librefang_types::event::EventTarget::Broadcast,
-                librefang_types::event::EventPayload::System(
-                    librefang_types::event::SystemEvent::TaskClaimed {
-                        task_id,
-                        claimed_by: resolved.clone(),
-                    },
-                ),
-            );
-            self.publish_event(event).await;
-        }
-
-        Ok(result)
-    }
-
-    async fn task_complete(
-        &self,
-        agent_id: &str,
-        task_id: &str,
-        result: &str,
-    ) -> Result<(), kernel_handle::KernelOpError> {
-        use kernel_handle::KernelOpError;
-        let resolved = match librefang_types::agent::AgentId::from_str(agent_id) {
-            Ok(_) => agent_id.to_string(),
-            Err(_) => match self.registry.find_by_name(agent_id) {
-                Some(entry) => entry.id.to_string(),
-                None => {
-                    return Err(KernelOpError::AgentNotFound(agent_id.to_string()));
-                }
-            },
-        };
-        self.memory
-            .task_complete(task_id, result)
-            .await
-            .map_err(|e| KernelOpError::Internal(format!("Task complete failed: {e}")))?;
-
-        let event = librefang_types::event::Event::new(
-            AgentId::new(), // system-originated
-            librefang_types::event::EventTarget::Broadcast,
-            librefang_types::event::EventPayload::System(
-                librefang_types::event::SystemEvent::TaskCompleted {
-                    task_id: task_id.to_string(),
-                    completed_by: resolved,
-                    result: result.to_string(),
-                },
-            ),
-        );
-        self.publish_event(event).await;
-
-        Ok(())
-    }
-
-    async fn task_list(
-        &self,
-        status: Option<&str>,
-    ) -> Result<Vec<serde_json::Value>, kernel_handle::KernelOpError> {
-        self.memory
-            .task_list(status)
-            .await
-            .map_err(|e| kernel_handle::KernelOpError::Internal(format!("Task list failed: {e}")))
-    }
-
-    async fn task_delete(&self, task_id: &str) -> Result<bool, kernel_handle::KernelOpError> {
-        self.memory
-            .task_delete(task_id)
-            .await
-            .map_err(|e| kernel_handle::KernelOpError::Internal(format!("Task delete failed: {e}")))
-    }
-
-    async fn task_retry(&self, task_id: &str) -> Result<bool, kernel_handle::KernelOpError> {
-        self.memory
-            .task_retry(task_id)
-            .await
-            .map_err(|e| kernel_handle::KernelOpError::Internal(format!("Task retry failed: {e}")))
-    }
-
-    async fn task_get(
-        &self,
-        task_id: &str,
-    ) -> Result<Option<serde_json::Value>, kernel_handle::KernelOpError> {
-        self.memory
-            .task_get(task_id)
-            .await
-            .map_err(|e| kernel_handle::KernelOpError::Internal(format!("Task get failed: {e}")))
-    }
-
-    async fn task_update_status(
-        &self,
-        task_id: &str,
-        new_status: &str,
-    ) -> Result<bool, kernel_handle::KernelOpError> {
-        self.memory
-            .task_update_status(task_id, new_status)
-            .await
-            .map_err(|e| {
-                kernel_handle::KernelOpError::Internal(format!("Task update status failed: {e}"))
-            })
-    }
-}
-
-#[async_trait::async_trait]
-impl kernel_handle::EventBus for LibreFangKernel {
-    async fn publish_event(
-        &self,
-        event_type: &str,
-        payload: serde_json::Value,
-    ) -> Result<(), kernel_handle::KernelOpError> {
-        let system_agent = AgentId::new();
-        // `?` lifts via `From<serde_json::Error>` on KernelOpError.
-        let payload_bytes =
-            serde_json::to_vec(&serde_json::json!({"type": event_type, "data": payload}))?;
-        let event = Event::new(
-            system_agent,
-            EventTarget::Broadcast,
-            EventPayload::Custom(payload_bytes),
-        );
-        LibreFangKernel::publish_event(self, event).await;
-        Ok(())
-    }
-}
-
-#[async_trait::async_trait]
-impl kernel_handle::KnowledgeGraph for LibreFangKernel {
-    async fn knowledge_add_entity(
-        &self,
-        entity: &librefang_types::memory::Entity,
-    ) -> Result<String, kernel_handle::KernelOpError> {
-        // The substrate owns the value (it moves into spawn_blocking).
-        // Clone here so the trait can take `&Entity` and avoid forcing
-        // every caller to give up ownership. See #3553.
-        self.memory.add_entity(entity.clone()).await.map_err(|e| {
-            kernel_handle::KernelOpError::Internal(format!("Knowledge add entity failed: {e}"))
-        })
-    }
-
-    async fn knowledge_add_relation(
-        &self,
-        relation: &librefang_types::memory::Relation,
-    ) -> Result<String, kernel_handle::KernelOpError> {
-        self.memory
-            .add_relation(relation.clone())
-            .await
-            .map_err(|e| {
-                kernel_handle::KernelOpError::Internal(format!(
-                    "Knowledge add relation failed: {e}"
-                ))
-            })
-    }
-
-    async fn knowledge_query(
-        &self,
-        pattern: librefang_types::memory::GraphPattern,
-    ) -> Result<Vec<librefang_types::memory::GraphMatch>, kernel_handle::KernelOpError> {
-        self.memory.query_graph(pattern).await.map_err(|e| {
-            kernel_handle::KernelOpError::Internal(format!("Knowledge query failed: {e}"))
-        })
-    }
-}
-
-#[async_trait::async_trait]
-impl kernel_handle::CronControl for LibreFangKernel {
-    async fn cron_create(
-        &self,
-        agent_id: &str,
-        job_json: serde_json::Value,
-    ) -> Result<String, kernel_handle::KernelOpError> {
-        use kernel_handle::KernelOpError;
-        use librefang_types::scheduler::{
-            CronAction, CronDelivery, CronDeliveryTarget, CronJob, CronJobId, CronSchedule,
-        };
-
-        let name = job_json["name"]
-            .as_str()
-            .ok_or_else(|| {
-                KernelOpError::InvalidInput("name: missing or not a string".to_string())
-            })?
-            .to_string();
-        let schedule: CronSchedule = serde_json::from_value(job_json["schedule"].clone())
-            .map_err(|e| KernelOpError::InvalidInput(format!("schedule: {e}")))?;
-        let action: CronAction = serde_json::from_value(job_json["action"].clone())
-            .map_err(|e| KernelOpError::InvalidInput(format!("action: {e}")))?;
-        let delivery: CronDelivery = if job_json["delivery"].is_object() {
-            serde_json::from_value(job_json["delivery"].clone())
-                .map_err(|e| KernelOpError::InvalidInput(format!("delivery: {e}")))?
-        } else {
-            // Default to LastChannel so cron jobs created by an agent in
-            // a channel context actually deliver their output back to
-            // that channel. The previous default (`None`) silently
-            // dropped every result and gave users no way to recover the
-            // originating channel without explicit `delivery` config.
-            // Issue #2338.
-            CronDelivery::LastChannel
-        };
-        // At-schedules are inherently single-execution; default one_shot=true for them
-        // so the job auto-deletes after firing instead of lingering as a zombie (#2808).
-        let is_at_schedule = matches!(schedule, CronSchedule::At { .. });
-        let one_shot = job_json["one_shot"].as_bool().unwrap_or(is_at_schedule);
-
-        let aid = librefang_types::agent::AgentId(
-            uuid::Uuid::parse_str(agent_id)
-                .map_err(|e| KernelOpError::InvalidInput(format!("agent_id: {e}")))?,
-        );
-
-        let session_mode: Option<librefang_types::agent::SessionMode> =
-            if job_json["session_mode"].is_string() {
-                serde_json::from_value(job_json["session_mode"].clone())
-                    .map_err(|e| KernelOpError::InvalidInput(format!("session_mode: {e}")))?
-            } else {
-                None
-            };
-
-        // Multi-destination fan-out targets. Optional; missing/null = empty.
-        // Validate each entry up front so a bad shape produces a clear error
-        // before the job is added (rather than failing silently at fire time).
-        let delivery_targets: Vec<CronDeliveryTarget> = if job_json["delivery_targets"].is_array() {
-            serde_json::from_value(job_json["delivery_targets"].clone())
-                .map_err(|e| KernelOpError::InvalidInput(format!("delivery_targets: {e}")))?
-        } else {
-            Vec::new()
-        };
-
-        let job = CronJob {
-            id: CronJobId::new(),
-            agent_id: aid,
-            name,
-            schedule,
-            action,
-            delivery,
-            delivery_targets,
-            peer_id: job_json["peer_id"].as_str().map(|s| s.to_string()),
-            session_mode,
-            enabled: true,
-            created_at: chrono::Utc::now(),
-            next_run: None,
-            last_run: None,
-        };
-
-        let id = self
-            .cron_scheduler
-            .add_job(job, one_shot)
-            .map_err(|e| KernelOpError::Internal(e.to_string()))?;
-
-        // Persist after adding
-        if let Err(e) = self.cron_scheduler.persist() {
-            tracing::warn!("Failed to persist cron jobs: {e}");
-        }
-
-        Ok(serde_json::json!({
-            "job_id": id.to_string(),
-            "status": "created"
-        })
-        .to_string())
-    }
-
-    async fn cron_list(
-        &self,
-        agent_id: &str,
-    ) -> Result<Vec<serde_json::Value>, kernel_handle::KernelOpError> {
-        use kernel_handle::KernelOpError;
-        let aid = librefang_types::agent::AgentId(
-            uuid::Uuid::parse_str(agent_id)
-                .map_err(|e| KernelOpError::InvalidInput(format!("agent_id: {e}")))?,
-        );
-        let jobs = self.cron_scheduler.list_jobs(aid);
-        let json_jobs: Vec<serde_json::Value> = jobs
-            .into_iter()
-            .map(|j| serde_json::to_value(&j).unwrap_or_default())
-            .collect();
-        Ok(json_jobs)
-    }
-
-    async fn cron_cancel(&self, job_id: &str) -> Result<(), kernel_handle::KernelOpError> {
-        use kernel_handle::KernelOpError;
-        let id = librefang_types::scheduler::CronJobId(
-            uuid::Uuid::parse_str(job_id)
-                .map_err(|e| KernelOpError::InvalidInput(format!("job_id: {e}")))?,
-        );
-        self.cron_scheduler
-            .remove_job(id)
-            .map_err(|e| KernelOpError::Internal(e.to_string()))?;
-
-        // Persist after removal
-        if let Err(e) = self.cron_scheduler.persist() {
-            tracing::warn!("Failed to persist cron jobs: {e}");
-        }
-
-        Ok(())
-    }
-}
-
-#[async_trait::async_trait]
-impl kernel_handle::HandsControl for LibreFangKernel {
-    async fn hand_list(&self) -> Result<Vec<serde_json::Value>, kernel_handle::KernelOpError> {
-        let defs = self.hand_registry.list_definitions();
-        let instances = self.hand_registry.list_instances();
-
-        let mut result = Vec::new();
-        for def in defs {
-            // Check if this hand has an active instance
-            let active_instance = instances.iter().find(|i| i.hand_id == def.id);
-            let (status, instance_id, agent_id) = match active_instance {
-                Some(inst) => (
-                    format!("{}", inst.status),
-                    Some(inst.instance_id.to_string()),
-                    inst.agent_id().map(|a: AgentId| a.to_string()),
-                ),
-                None => ("available".to_string(), None, None),
-            };
-
-            let mut entry = serde_json::json!({
-                "id": def.id,
-                "name": def.name,
-                "icon": def.icon,
-                "category": format!("{:?}", def.category),
-                "description": def.description,
-                "status": status,
-                "tools": def.tools,
-            });
-            if let Some(iid) = instance_id {
-                entry["instance_id"] = serde_json::json!(iid);
-            }
-            if let Some(aid) = agent_id {
-                entry["agent_id"] = serde_json::json!(aid);
-            }
-            result.push(entry);
-        }
-        Ok(result)
-    }
-
-    async fn hand_install(
-        &self,
-        toml_content: &str,
-        skill_content: &str,
-    ) -> Result<serde_json::Value, kernel_handle::KernelOpError> {
-        let def = self
-            .hand_registry
-            .install_from_content_persisted(&self.home_dir_boot, toml_content, skill_content)
-            .map_err(|e| format!("{e}"))?;
-        router::invalidate_hand_route_cache();
-
-        Ok(serde_json::json!({
-            "id": def.id,
-            "name": def.name,
-            "description": def.description,
-            "category": format!("{:?}", def.category),
-        }))
-    }
-
-    async fn hand_activate(
-        &self,
-        hand_id: &str,
-        config: std::collections::HashMap<String, serde_json::Value>,
-    ) -> Result<serde_json::Value, kernel_handle::KernelOpError> {
-        let instance = self
-            .activate_hand(hand_id, config)
-            .map_err(|e| format!("{e}"))?;
-
-        Ok(serde_json::json!({
-            "instance_id": instance.instance_id.to_string(),
-            "hand_id": instance.hand_id,
-            "agent_name": instance.agent_name(),
-            "agent_id": instance.agent_id().map(|a| a.to_string()),
-            "status": format!("{}", instance.status),
-        }))
-    }
-
-    async fn hand_status(
-        &self,
-        hand_id: &str,
-    ) -> Result<serde_json::Value, kernel_handle::KernelOpError> {
-        let instances = self.hand_registry.list_instances();
-        let instance = instances
-            .iter()
-            .find(|i| i.hand_id == hand_id)
-            .ok_or_else(|| format!("No active instance found for hand '{hand_id}'"))?;
-
-        let def = self.hand_registry.get_definition(hand_id);
-        let def_name = def.as_ref().map(|d| d.name.clone()).unwrap_or_default();
-        let def_icon = def.as_ref().map(|d| d.icon.clone()).unwrap_or_default();
-
-        Ok(serde_json::json!({
-            "hand_id": hand_id,
-            "name": def_name,
-            "icon": def_icon,
-            "instance_id": instance.instance_id.to_string(),
-            "status": format!("{}", instance.status),
-            "agent_id": instance.agent_id().map(|a| a.to_string()),
-            "agent_name": instance.agent_name(),
-            "activated_at": instance.activated_at.to_rfc3339(),
-            "updated_at": instance.updated_at.to_rfc3339(),
-        }))
-    }
-
-    async fn hand_deactivate(&self, instance_id: &str) -> Result<(), kernel_handle::KernelOpError> {
-        use kernel_handle::KernelOpError;
-        let uuid = uuid::Uuid::parse_str(instance_id)
-            .map_err(|e| KernelOpError::InvalidInput(format!("instance_id: {e}")))?;
-        self.deactivate_hand(uuid)
-            .map_err(|e| KernelOpError::Internal(e.to_string()))
-    }
-}
-
-#[async_trait::async_trait]
-impl kernel_handle::ApprovalGate for LibreFangKernel {
-    fn requires_approval(&self, tool_name: &str) -> bool {
-        self.approval_manager.requires_approval(tool_name)
-    }
-
-    fn requires_approval_with_context(
-        &self,
-        tool_name: &str,
-        sender_id: Option<&str>,
-        channel: Option<&str>,
-    ) -> bool {
-        self.approval_manager
-            .requires_approval_with_context(tool_name, sender_id, channel)
-    }
-
-    fn is_tool_denied_with_context(
-        &self,
-        tool_name: &str,
-        sender_id: Option<&str>,
-        channel: Option<&str>,
-    ) -> bool {
-        self.approval_manager
-            .is_tool_denied_with_context(tool_name, sender_id, channel)
-    }
-
-    fn resolve_user_tool_decision(
-        &self,
-        tool_name: &str,
-        sender_id: Option<&str>,
-        channel: Option<&str>,
-    ) -> librefang_types::user_policy::UserToolGate {
-        // The synthetic `"cron"` and `"autonomous"` channels are the only
-        // two the kernel treats as system-internal. Both are synthesised
-        // by the kernel itself for daemon-driven calls that have no
-        // user-facing sender:
-        //   - `"cron"` — `kernel/mod.rs::start_periodic_loops` cron tick
-        //     (~line 11950) for `[[cron_jobs]]` fires.
-        //   - `"autonomous"` — `start_continuous_autonomous_loop`
-        //     (~line 12412) for autonomous-tick prompts on agents whose
-        //     manifest declares `[autonomous]`.
-        // Both fan out the agent's own loop with a synthetic
-        // `SenderContext { channel: "cron" | "autonomous" }`. Issue #3243
-        // tracks the autonomous case: without this carve-out, every
-        // autonomous tool call falls into `guest_gate` → NeedsApproval
-        // and floods the approval queue when RBAC is enabled.
-        //
-        // Earlier drafts also matched `"system"` / `"internal"` and
-        // treated `(None, None)` as system, but neither sentinel is
-        // synthesised anywhere in the codebase, and the `(None, None)`
-        // shortcut silently re-opened the H7 fail-open at the trait
-        // boundary the AuthManager unit tests were written to close
-        // (PR #3205 review item #1). Both have been removed: an
-        // unattributed inbound now goes through the guest gate so
-        // RBAC fails closed end-to-end.
-        let system_call = matches!(
-            channel,
-            Some(c) if c == SYSTEM_CHANNEL_CRON || c == SYSTEM_CHANNEL_AUTONOMOUS
-        );
-        self.auth
-            .resolve_user_tool_decision(tool_name, sender_id, channel, system_call)
-    }
-
-    async fn request_approval(
-        &self,
-        agent_id: &str,
-        tool_name: &str,
-        action_summary: &str,
-        session_id: Option<&str>,
-    ) -> Result<librefang_types::approval::ApprovalDecision, kernel_handle::KernelOpError> {
-        use librefang_types::approval::{ApprovalDecision, ApprovalRequest as TypedRequest};
-
-        // Hand agents are curated trusted packages — auto-approve tool execution.
-        // Check if this agent has a "hand:" tag indicating it was spawned by activate_hand().
-        if let Ok(aid) = agent_id.parse::<AgentId>() {
-            if let Some(entry) = self.registry.get(aid) {
-                if entry.tags.iter().any(|t| t.starts_with("hand:")) {
-                    info!(agent_id, tool_name, "Auto-approved for hand agent");
-                    return Ok(ApprovalDecision::Approved);
-                }
-            }
-        }
-
-        let policy = self.approval_manager.policy();
-        let risk_level = crate::approval::ApprovalManager::classify_risk(tool_name);
-        let agent_display = self.approval_agent_display(agent_id);
-        let description = format!("Agent {} requests to execute {}", agent_display, tool_name);
-        let request_id = uuid::Uuid::new_v4();
-        let req = TypedRequest {
-            id: request_id,
-            agent_id: agent_id.to_string(),
-            tool_name: tool_name.to_string(),
-            description: description.clone(),
-            action_summary: action_summary
-                .chars()
-                .take(librefang_types::approval::MAX_ACTION_SUMMARY_LEN)
-                .collect(),
-            risk_level,
-            requested_at: chrono::Utc::now(),
-            timeout_secs: policy.timeout_secs,
-            sender_id: None,
-            channel: None,
-            route_to: Vec::new(),
-            escalation_count: 0,
-            session_id: session_id.map(|s| s.to_string()),
-        };
-
-        // Publish an ApprovalRequested event so channel adapters can notify users
-        {
-            use librefang_types::event::{
-                ApprovalRequestedEvent, Event, EventPayload, EventTarget,
-            };
-            let event = Event::new(
-                agent_id.parse().unwrap_or_default(),
-                EventTarget::System,
-                EventPayload::ApprovalRequested(ApprovalRequestedEvent {
-                    request_id: request_id.to_string(),
-                    agent_id: agent_id.to_string(),
-                    tool_name: tool_name.to_string(),
-                    description: description.clone(),
-                    risk_level: format!("{:?}", risk_level),
-                }),
-            );
-            self.event_bus.publish(event).await;
-        }
-
-        // Push approval notification to configured channels.
-        // Resolution order: per-request route_to > policy routing rules > per-agent rules > global defaults.
-        {
-            use librefang_types::capability::glob_matches;
-
-            let cfg = self.config.load_full();
-            let policy = self.approval_manager.policy();
-            let targets: Vec<librefang_types::approval::NotificationTarget> =
-                if !req.route_to.is_empty() {
-                    // Highest priority: explicitly routed targets on the request itself
-                    req.route_to.clone()
-                } else {
-                    // Check policy routing rules (match tool_pattern)
-                    let routed: Vec<librefang_types::approval::NotificationTarget> = policy
-                        .routing
-                        .iter()
-                        .filter(|r| glob_matches(&r.tool_pattern, tool_name))
-                        .flat_map(|r| r.route_to.clone())
-                        .collect();
-                    if !routed.is_empty() {
-                        routed
-                    } else {
-                        // Check per-agent notification rules
-                        let agent_routed: Vec<librefang_types::approval::NotificationTarget> = cfg
-                            .notification
-                            .agent_rules
-                            .iter()
-                            .filter(|rule| {
-                                glob_matches(&rule.agent_pattern, agent_id)
-                                    && rule.events.iter().any(|e| e == "approval_requested")
-                            })
-                            .flat_map(|rule| rule.channels.clone())
-                            .collect();
-                        if !agent_routed.is_empty() {
-                            agent_routed
-                        } else {
-                            // Fallback: global approval_channels
-                            cfg.notification.approval_channels.clone()
-                        }
-                    }
-                };
-
-            let msg = format!(
-                "{} Approval needed: agent {} wants to run `{}` — {}",
-                risk_level.emoji(),
-                agent_display,
-                tool_name,
-                description,
-            );
-            let req_id_str = request_id.to_string();
-            for target in &targets {
-                self.push_approval_interactive(target, &msg, &req_id_str)
-                    .await;
-            }
-        }
-
-        let decision = self.approval_manager.request_approval(req).await;
-
-        // Publish resolved event so channel adapters can notify outcome
-        {
-            use librefang_types::event::{ApprovalResolvedEvent, Event, EventPayload, EventTarget};
-            let event = Event::new(
-                agent_id.parse().unwrap_or_default(),
-                EventTarget::System,
-                EventPayload::ApprovalResolved(ApprovalResolvedEvent {
-                    request_id: request_id.to_string(),
-                    agent_id: agent_id.to_string(),
-                    tool_name: tool_name.to_string(),
-                    decision: decision.as_str().to_string(),
-                    decided_by: None,
-                }),
-            );
-            self.event_bus.publish(event).await;
-        }
-
-        Ok(decision)
-    }
-
-    async fn submit_tool_approval(
-        &self,
-        agent_id: &str,
-        tool_name: &str,
-        action_summary: &str,
-        deferred: librefang_types::tool::DeferredToolExecution,
-        session_id: Option<&str>,
-    ) -> Result<ToolApprovalSubmission, kernel_handle::KernelOpError> {
-        use librefang_types::approval::ApprovalRequest as TypedRequest;
-
-        // Hand agents are curated trusted packages — auto-approve for non-blocking execution.
-        // EXCEPTION (RBAC M3, #3054): when the per-user policy demanded approval
-        // (`force_human=true`), the carve-out MUST NOT fire — otherwise a Viewer/User
-        // chatting with a hand-tagged agent silently inherits the agent's full
-        // tool surface, defeating user-level RBAC entirely.
-        if !deferred.force_human {
-            if let Ok(aid) = agent_id.parse::<AgentId>() {
-                if let Some(entry) = self.registry.get(aid) {
-                    if entry.tags.iter().any(|t| t.starts_with("hand:")) {
-                        info!(
-                            agent_id,
-                            tool_name, "Auto-approved for hand agent (non-blocking)"
-                        );
-                        return Ok(ToolApprovalSubmission::AutoApproved);
-                    }
-                }
-            }
-        } else {
-            debug!(
-                agent_id,
-                tool_name, "Hand-agent auto-approval skipped because user policy demanded approval"
-            );
-        }
-
-        let policy = self.approval_manager.policy();
-        let risk_level = crate::approval::ApprovalManager::classify_risk(tool_name);
-        let agent_display = self.approval_agent_display(agent_id);
-        let description = format!("Agent {} requests to execute {}", agent_display, tool_name);
-        let request_id = uuid::Uuid::new_v4();
-        let req = TypedRequest {
-            id: request_id,
-            agent_id: agent_id.to_string(),
-            tool_name: tool_name.to_string(),
-            description: description.clone(),
-            action_summary: action_summary
-                .chars()
-                .take(librefang_types::approval::MAX_ACTION_SUMMARY_LEN)
-                .collect(),
-            risk_level,
-            requested_at: chrono::Utc::now(),
-            timeout_secs: policy.timeout_secs,
-            sender_id: None,
-            channel: None,
-            route_to: Vec::new(),
-            escalation_count: 0,
-            session_id: session_id.map(|s| s.to_string()),
-        };
-
-        self.approval_manager
-            .submit_request(req.clone(), deferred)
-            .map_err(|e| e.to_string())?;
-
-        // Publish event + push notification (same as blocking path)
-        {
-            use librefang_types::event::{
-                ApprovalRequestedEvent, Event, EventPayload, EventTarget,
-            };
-            let event = Event::new(
-                agent_id.parse().unwrap_or_default(),
-                EventTarget::System,
-                EventPayload::ApprovalRequested(ApprovalRequestedEvent {
-                    request_id: request_id.to_string(),
-                    agent_id: agent_id.to_string(),
-                    tool_name: tool_name.to_string(),
-                    description: description.clone(),
-                    risk_level: format!("{:?}", risk_level),
-                }),
-            );
-            self.event_bus.publish(event).await;
-        }
-        {
-            use librefang_types::capability::glob_matches;
-            let cfg = self.config.load_full();
-            let targets: Vec<librefang_types::approval::NotificationTarget> = {
-                let routed: Vec<_> = policy
-                    .routing
-                    .iter()
-                    .filter(|r| glob_matches(&r.tool_pattern, tool_name))
-                    .flat_map(|r| r.route_to.clone())
-                    .collect();
-                if !routed.is_empty() {
-                    routed
-                } else {
-                    let agent_routed: Vec<_> = cfg
-                        .notification
-                        .agent_rules
-                        .iter()
-                        .filter(|rule| {
-                            glob_matches(&rule.agent_pattern, agent_id)
-                                && rule.events.iter().any(|e| e == "approval_requested")
-                        })
-                        .flat_map(|rule| rule.channels.clone())
-                        .collect();
-                    if !agent_routed.is_empty() {
-                        agent_routed
-                    } else {
-                        cfg.notification.approval_channels.clone()
-                    }
-                }
-            };
-            let msg = format!(
-                "{} Approval needed: agent {} wants to run `{}` — {}",
-                risk_level.emoji(),
-                agent_display,
-                tool_name,
-                description,
-            );
-            let req_id_str = request_id.to_string();
-            for target in &targets {
-                self.push_approval_interactive(target, &msg, &req_id_str)
-                    .await;
-            }
-        }
-
-        Ok(ToolApprovalSubmission::Pending { request_id })
-    }
-
-    async fn resolve_tool_approval(
-        &self,
-        request_id: uuid::Uuid,
-        decision: librefang_types::approval::ApprovalDecision,
-        decided_by: Option<String>,
-        totp_verified: bool,
-        user_id: Option<&str>,
-    ) -> Result<
-        (
-            librefang_types::approval::ApprovalResponse,
-            Option<librefang_types::tool::DeferredToolExecution>,
-        ),
-        kernel_handle::KernelOpError,
-    > {
-        // #3541 follow-up: classify the missing-id case as
-        // `KernelOpError::AgentNotFound` / `Internal` so the API
-        // boundary surfaces 404 via the typed mapping. The underlying
-        // `ApprovalManager::resolve` still returns `String` (typing it
-        // is left to a separate ApprovalManager refactor); the substring
-        // check is scoped to the manager's exact "not found or expired"
-        // wording. All other error wordings flow through `Internal`.
-        let (response, deferred) = self
-            .approval_manager
-            .resolve(request_id, decision, decided_by, totp_verified, user_id)
-            .map_err(|msg| {
-                if msg.contains("not found") {
-                    kernel_handle::KernelOpError::AgentNotFound(request_id.to_string())
-                } else {
-                    kernel_handle::KernelOpError::Internal(msg)
-                }
-            })?;
-
-        // Deferred approval execution resumes in the background so API callers do
-        // not block on slow tools.
-        if let Some(ref def) = deferred {
-            let decision_clone = response.decision.clone();
-            let kernel = Arc::clone(
-                self.self_handle
-                    .get()
-                    .and_then(|w| w.upgrade())
-                    .as_ref()
-                    .ok_or_else(|| "Kernel self-handle unavailable".to_string())?,
-            );
-            let deferred_clone = def.clone();
-            spawn_logged("approval_resolution", async move {
-                kernel
-                    .handle_approval_resolution(request_id, decision_clone, deferred_clone)
-                    .await;
-            });
-        }
-
-        Ok((response, deferred))
-    }
-
-    fn get_approval_status(
-        &self,
-        request_id: uuid::Uuid,
-    ) -> Result<Option<librefang_types::approval::ApprovalDecision>, kernel_handle::KernelOpError>
-    {
-        // If still pending, no decision yet.
-        if self.approval_manager.get_pending(request_id).is_some() {
-            return Ok(None);
-        }
-        // Check recent resolved records.
-        let recent = self.approval_manager.list_recent(200);
-        for record in &recent {
-            if record.request.id == request_id {
-                return Ok(Some(record.decision.clone()));
-            }
-        }
-        Ok(None)
-    }
-}
-
-impl kernel_handle::A2ARegistry for LibreFangKernel {
-    fn list_a2a_agents(&self) -> Vec<(String, String)> {
-        let agents = self
-            .a2a_external_agents
-            .lock()
-            .unwrap_or_else(|e| e.into_inner());
-        // Return (name, key) pairs where `key` is the trust-list key
-        // (first tuple element), not `card.url`. The card's self-declared
-        // url is `<base>/a2a` while the trust gate at /api/a2a/send and
-        // tool_a2a_send compare against the canonicalized base URL. Using
-        // `card.url` here would silently mismatch the gate and break every
-        // statically-seeded entry. (Bug #3786)
-        agents
-            .iter()
-            .map(|(key, card)| (card.name.clone(), key.clone()))
-            .collect()
-    }
-
-    fn get_a2a_agent_url(&self, name: &str) -> Option<String> {
-        let agents = self
-            .a2a_external_agents
-            .lock()
-            .unwrap_or_else(|e| e.into_inner());
-        let name_lower = name.to_lowercase();
-        // See list_a2a_agents — return the trust-list key, not card.url,
-        // so callers get a URL that the gate will accept.
-        agents
-            .iter()
-            .find(|(_, card)| card.name.to_lowercase() == name_lower)
-            .map(|(key, _)| key.clone())
-    }
-}
-
-#[async_trait::async_trait]
-impl kernel_handle::ChannelSender for LibreFangKernel {
-    async fn send_channel_message(
-        &self,
-        channel: &str,
-        recipient: &str,
-        message: &str,
-        thread_id: Option<&str>,
-        account_id: Option<&str>,
-    ) -> Result<String, kernel_handle::KernelOpError> {
-        let cfg = self.config.load_full();
-        let lookup_key = account_id
-            .filter(|s| !s.is_empty())
-            .map(|aid| format!("{channel}:{aid}"))
-            .unwrap_or_else(|| channel.to_string());
-        let adapter = self
-            .channel_adapters
-            .get(&lookup_key)
-            .ok_or_else(|| {
-                let available: Vec<String> = self
-                    .channel_adapters
-                    .iter()
-                    .map(|e| e.key().clone())
-                    .collect();
-                match account_id.filter(|s| !s.is_empty()) {
-                    Some(aid) => format!(
-                        "Channel '{}' with account_id '{}' not found. Available: {:?}",
-                        channel, aid, available
-                    ),
-                    None => format!(
-                        "Channel '{}' not found. Available channels: {:?}",
-                        channel, available
-                    ),
-                }
-            })?
-            .clone();
-
-        let user = librefang_channels::types::ChannelUser {
-            platform_id: recipient.to_string(),
-            display_name: recipient.to_string(),
-            librefang_user: None,
-        };
-
-        let default_format =
-            librefang_channels::formatter::default_output_format_for_channel(channel);
-        let formatted = if channel == "wecom" {
-            let output_format = cfg
-                .channels
-                .wecom
-                .as_ref()
-                .and_then(|c| c.overrides.output_format)
-                .unwrap_or(default_format);
-            librefang_channels::formatter::format_for_wecom(message, output_format)
-        } else {
-            librefang_channels::formatter::format_for_channel(message, default_format)
-        };
-
-        let content = librefang_channels::types::ChannelContent::Text(formatted);
-
-        if let Some(tid) = thread_id {
-            adapter
-                .send_in_thread(&user, content, tid)
-                .await
-                .map_err(|e| format!("Channel send failed: {e}"))?;
-        } else {
-            adapter
-                .send(&user, content)
-                .await
-                .map_err(|e| format!("Channel send failed: {e}"))?;
-        }
-
-        Ok(format!("Message sent to {} via {}", recipient, channel))
-    }
-
-    async fn send_channel_media(
-        &self,
-        channel: &str,
-        recipient: &str,
-        media_type: &str,
-        media_url: &str,
-        caption: Option<&str>,
-        filename: Option<&str>,
-        thread_id: Option<&str>,
-        account_id: Option<&str>,
-    ) -> Result<String, kernel_handle::KernelOpError> {
-        let lookup_key = account_id
-            .filter(|s| !s.is_empty())
-            .map(|aid| format!("{channel}:{aid}"))
-            .unwrap_or_else(|| channel.to_string());
-        let adapter = self
-            .channel_adapters
-            .get(&lookup_key)
-            .ok_or_else(|| {
-                let available: Vec<String> = self
-                    .channel_adapters
-                    .iter()
-                    .map(|e| e.key().clone())
-                    .collect();
-                match account_id.filter(|s| !s.is_empty()) {
-                    Some(aid) => format!(
-                        "Channel '{}' with account_id '{}' not found. Available: {:?}",
-                        channel, aid, available
-                    ),
-                    None => format!(
-                        "Channel '{}' not found. Available channels: {:?}",
-                        channel, available
-                    ),
-                }
-            })?
-            .clone();
-
-        let user = librefang_channels::types::ChannelUser {
-            platform_id: recipient.to_string(),
-            display_name: recipient.to_string(),
-            librefang_user: None,
-        };
-
-        let content = match media_type {
-            "image" => librefang_channels::types::ChannelContent::Image {
-                url: media_url.to_string(),
-                caption: caption.map(|s| s.to_string()),
-                mime_type: None,
-            },
-            "file" => librefang_channels::types::ChannelContent::File {
-                url: media_url.to_string(),
-                filename: filename.unwrap_or("file").to_string(),
-            },
-            _ => {
-                return Err(kernel_handle::KernelOpError::InvalidInput(format!(
-                    "media_type: Unsupported media type: '{media_type}'. Use 'image' or 'file'."
-                )));
-            }
-        };
-
-        if let Some(tid) = thread_id {
-            adapter
-                .send_in_thread(&user, content, tid)
-                .await
-                .map_err(|e| format!("Channel media send failed: {e}"))?;
-        } else {
-            adapter
-                .send(&user, content)
-                .await
-                .map_err(|e| format!("Channel media send failed: {e}"))?;
-        }
-
-        Ok(format!(
-            "{} sent to {} via {}",
-            media_type, recipient, channel
-        ))
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    async fn send_channel_file_data(
-        &self,
-        channel: &str,
-        recipient: &str,
-        data: bytes::Bytes,
-        filename: &str,
-        mime_type: &str,
-        thread_id: Option<&str>,
-        account_id: Option<&str>,
-    ) -> Result<String, kernel_handle::KernelOpError> {
-        let lookup_key = account_id
-            .filter(|s| !s.is_empty())
-            .map(|aid| format!("{channel}:{aid}"))
-            .unwrap_or_else(|| channel.to_string());
-        let adapter = self
-            .channel_adapters
-            .get(&lookup_key)
-            .ok_or_else(|| {
-                let available: Vec<String> = self
-                    .channel_adapters
-                    .iter()
-                    .map(|e| e.key().clone())
-                    .collect();
-                match account_id.filter(|s| !s.is_empty()) {
-                    Some(aid) => format!(
-                        "Channel '{}' with account_id '{}' not found. Available: {:?}",
-                        channel, aid, available
-                    ),
-                    None => format!(
-                        "Channel '{}' not found. Available channels: {:?}",
-                        channel, available
-                    ),
-                }
-            })?
-            .clone();
-
-        let user = librefang_channels::types::ChannelUser {
-            platform_id: recipient.to_string(),
-            display_name: recipient.to_string(),
-            librefang_user: None,
-        };
-
-        // `ChannelContent::FileData` still carries `Vec<u8>` (changing it
-        // is out of scope for #3553 — that's a follow-up that touches
-        // every channel adapter). `Vec::from(Bytes)` is O(1) when the
-        // Bytes uniquely owns its allocation, which is the common case
-        // here (caller built it via `Bytes::from(vec)` straight from
-        // `tokio::fs::read`).
-        let content = librefang_channels::types::ChannelContent::FileData {
-            data: Vec::from(data),
-            filename: filename.to_string(),
-            mime_type: mime_type.to_string(),
-        };
-
-        if let Some(tid) = thread_id {
-            adapter
-                .send_in_thread(&user, content, tid)
-                .await
-                .map_err(|e| format!("Channel file send failed: {e}"))?;
-        } else {
-            adapter
-                .send(&user, content)
-                .await
-                .map_err(|e| format!("Channel file send failed: {e}"))?;
-        }
-
-        Ok(format!(
-            "File '{}' sent to {} via {}",
-            filename, recipient, channel
-        ))
-    }
-
-    async fn send_channel_poll(
-        &self,
-        channel: &str,
-        recipient: &str,
-        question: &str,
-        options: &[String],
-        is_quiz: bool,
-        correct_option_id: Option<u8>,
-        explanation: Option<&str>,
-        account_id: Option<&str>,
-    ) -> Result<(), kernel_handle::KernelOpError> {
-        let lookup_key = account_id
-            .filter(|s| !s.is_empty())
-            .map(|aid| format!("{channel}:{aid}"))
-            .unwrap_or_else(|| channel.to_string());
-        let adapter = self
-            .channel_adapters
-            .get(&lookup_key)
-            .ok_or_else(|| match account_id.filter(|s| !s.is_empty()) {
-                Some(aid) => {
-                    format!("Channel adapter '{channel}' with account_id '{aid}' not found")
-                }
-                None => format!("Channel adapter '{channel}' not found"),
-            })?
-            .clone();
-
-        let user = librefang_channels::types::ChannelUser {
-            platform_id: recipient.to_string(),
-            display_name: recipient.to_string(),
-            librefang_user: None,
-        };
-
-        let content = librefang_channels::types::ChannelContent::Poll {
-            question: question.to_string(),
-            options: options.to_vec(),
-            is_quiz,
-            correct_option_id,
-            explanation: explanation.map(|s| s.to_string()),
-        };
-
-        adapter
-            .send(&user, content)
-            .await
-            .map_err(|e| format!("Channel poll send failed: {e}"))?;
-
-        Ok(())
-    }
-
-    fn roster_upsert(
-        &self,
-        channel: &str,
-        chat_id: &str,
-        user_id: &str,
-        display_name: &str,
-        username: Option<&str>,
-    ) -> Result<(), kernel_handle::KernelOpError> {
-        self.memory
-            .roster()
-            .upsert(channel, chat_id, user_id, display_name, username);
-        Ok(())
-    }
-
-    fn roster_members(
-        &self,
-        channel: &str,
-        chat_id: &str,
-    ) -> Result<Vec<serde_json::Value>, kernel_handle::KernelOpError> {
-        let members = self.memory.roster().members(channel, chat_id);
-        Ok(members
-            .into_iter()
-            .map(|(user_id, display_name, username)| {
-                serde_json::json!({
-                    "user_id": user_id,
-                    "display_name": display_name,
-                    "username": username,
-                })
-            })
-            .collect())
-    }
-
-    fn roster_remove_member(
-        &self,
-        channel: &str,
-        chat_id: &str,
-        user_id: &str,
-    ) -> Result<(), kernel_handle::KernelOpError> {
-        self.memory
-            .roster()
-            .remove_member(channel, chat_id, user_id);
-        Ok(())
-    }
-}
-
-impl kernel_handle::PromptStore for LibreFangKernel {
-    fn get_running_experiment(
-        &self,
-        agent_id: &str,
-    ) -> Result<Option<librefang_types::agent::PromptExperiment>, kernel_handle::KernelOpError>
-    {
-        let cfg = self.config.load();
-        if !cfg.prompt_intelligence.enabled {
-            return Ok(None);
-        }
-        let id: AgentId = agent_id.parse().map_err(|e| {
-            kernel_handle::KernelOpError::Internal(format!("Invalid agent ID: {e}"))
-        })?;
-        let store = self
-            .prompt_store
-            .get()
-            .ok_or("Prompt store not initialized")?;
-        store.get_running_experiment(id).map_err(|e| {
-            kernel_handle::KernelOpError::Internal(format!("Failed to get experiment: {e}"))
-        })
-    }
-
-    fn record_experiment_request(
-        &self,
-        experiment_id: &str,
-        variant_id: &str,
-        latency_ms: u64,
-        cost_usd: f64,
-        success: bool,
-    ) -> Result<(), kernel_handle::KernelOpError> {
-        let exp_id: uuid::Uuid = experiment_id.parse().map_err(|e| {
-            kernel_handle::KernelOpError::Internal(format!("Invalid experiment ID: {e}"))
-        })?;
-        let var_id: uuid::Uuid = variant_id.parse().map_err(|e| {
-            kernel_handle::KernelOpError::Internal(format!("Invalid variant ID: {e}"))
-        })?;
-        let store = self
-            .prompt_store
-            .get()
-            .ok_or(kernel_handle::KernelOpError::unavailable("Prompt store"))?;
-        store
-            .record_request(exp_id, var_id, latency_ms, cost_usd, success)
-            .map_err(|e| {
-                kernel_handle::KernelOpError::Internal(format!("Failed to record request: {e}"))
-            })
-    }
-
-    fn get_prompt_version(
-        &self,
-        version_id: &str,
-    ) -> Result<Option<librefang_types::agent::PromptVersion>, kernel_handle::KernelOpError> {
-        let id: uuid::Uuid = version_id.parse().map_err(|e| {
-            kernel_handle::KernelOpError::Internal(format!("Invalid version ID: {e}"))
-        })?;
-        let store = self
-            .prompt_store
-            .get()
-            .ok_or(kernel_handle::KernelOpError::unavailable("Prompt store"))?;
-        store.get_version(id).map_err(|e| {
-            kernel_handle::KernelOpError::Internal(format!("Failed to get version: {e}"))
-        })
-    }
-
-    fn list_prompt_versions(
-        &self,
-        agent_id: librefang_types::agent::AgentId,
-    ) -> Result<Vec<librefang_types::agent::PromptVersion>, kernel_handle::KernelOpError> {
-        let store = self
-            .prompt_store
-            .get()
-            .ok_or("Prompt store not initialized")?;
-        store.list_versions(agent_id).map_err(|e| {
-            kernel_handle::KernelOpError::Internal(format!("Failed to list versions: {e}"))
-        })
-    }
-
-    fn create_prompt_version(
-        &self,
-        version: &librefang_types::agent::PromptVersion,
-    ) -> Result<(), kernel_handle::KernelOpError> {
-        let cfg = self.config.load();
-        let store = self
-            .prompt_store
-            .get()
-            .ok_or(kernel_handle::KernelOpError::unavailable("Prompt store"))?;
-        let agent_id = version.agent_id;
-        // Clone here — the store owns the value. Trade-off accepted by
-        // #3553: callers (API handlers) no longer have to clone first.
-        store.create_version(version.clone()).map_err(|e| {
-            kernel_handle::KernelOpError::Internal(format!("Failed to create version: {e}"))
-        })?;
-        // Prune old versions if over the configured limit
-        let max = cfg.prompt_intelligence.max_versions_per_agent;
-        let _ = store.prune_old_versions(agent_id, max);
-        Ok(())
-    }
-
-    fn delete_prompt_version(&self, version_id: &str) -> Result<(), kernel_handle::KernelOpError> {
-        let id: uuid::Uuid = version_id.parse().map_err(|e| {
-            kernel_handle::KernelOpError::Internal(format!("Invalid version ID: {e}"))
-        })?;
-        let store = self
-            .prompt_store
-            .get()
-            .ok_or("Prompt store not initialized")?;
-        store.delete_version(id).map_err(|e| {
-            kernel_handle::KernelOpError::Internal(format!("Failed to delete version: {e}"))
-        })
-    }
-
-    fn set_active_prompt_version(
-        &self,
-        version_id: &str,
-        agent_id: &str,
-    ) -> Result<(), kernel_handle::KernelOpError> {
-        let id: uuid::Uuid = version_id.parse().map_err(|e| {
-            kernel_handle::KernelOpError::Internal(format!("Invalid version ID: {e}"))
-        })?;
-        let agent: librefang_types::agent::AgentId = agent_id.parse().map_err(|e| {
-            kernel_handle::KernelOpError::Internal(format!("Invalid agent ID: {e}"))
-        })?;
-        let store = self
-            .prompt_store
-            .get()
-            .ok_or("Prompt store not initialized")?;
-        store.set_active_version(id, agent).map_err(|e| {
-            kernel_handle::KernelOpError::Internal(format!("Failed to set active version: {e}"))
-        })
-    }
-
-    fn list_experiments(
-        &self,
-        agent_id: librefang_types::agent::AgentId,
-    ) -> Result<Vec<librefang_types::agent::PromptExperiment>, kernel_handle::KernelOpError> {
-        let store = self
-            .prompt_store
-            .get()
-            .ok_or("Prompt store not initialized")?;
-        store.list_experiments(agent_id).map_err(|e| {
-            kernel_handle::KernelOpError::Internal(format!("Failed to list experiments: {e}"))
-        })
-    }
-
-    fn create_experiment(
-        &self,
-        experiment: &librefang_types::agent::PromptExperiment,
-    ) -> Result<(), kernel_handle::KernelOpError> {
-        let store = self
-            .prompt_store
-            .get()
-            .ok_or(kernel_handle::KernelOpError::unavailable("Prompt store"))?;
-        // Clone here — the store owns the value. See #3553.
-        store.create_experiment(experiment.clone()).map_err(|e| {
-            kernel_handle::KernelOpError::Internal(format!("Failed to create experiment: {e}"))
-        })
-    }
-
-    fn get_experiment(
-        &self,
-        experiment_id: &str,
-    ) -> Result<Option<librefang_types::agent::PromptExperiment>, kernel_handle::KernelOpError>
-    {
-        let id: uuid::Uuid = experiment_id.parse().map_err(|e| {
-            kernel_handle::KernelOpError::Internal(format!("Invalid experiment ID: {e}"))
-        })?;
-        let store = self
-            .prompt_store
-            .get()
-            .ok_or("Prompt store not initialized")?;
-        store.get_experiment(id).map_err(|e| {
-            kernel_handle::KernelOpError::Internal(format!("Failed to get experiment: {e}"))
-        })
-    }
-
-    fn update_experiment_status(
-        &self,
-        experiment_id: &str,
-        status: librefang_types::agent::ExperimentStatus,
-    ) -> Result<(), kernel_handle::KernelOpError> {
-        let id: uuid::Uuid = experiment_id.parse().map_err(|e| {
-            kernel_handle::KernelOpError::Internal(format!("Invalid experiment ID: {e}"))
-        })?;
-        let store = self
-            .prompt_store
-            .get()
-            .ok_or("Prompt store not initialized")?;
-        store.update_experiment_status(id, status).map_err(|e| {
-            kernel_handle::KernelOpError::Internal(format!(
-                "Failed to update experiment status: {e}"
-            ))
-        })?;
-
-        // When completing an experiment, auto-activate the winning variant's prompt version
-        if status == librefang_types::agent::ExperimentStatus::Completed {
-            let metrics = store.get_experiment_metrics(id).map_err(|e| {
-                kernel_handle::KernelOpError::Internal(format!(
-                    "Failed to get experiment metrics: {e}"
-                ))
-            })?;
-            if let Some(winner) = metrics.iter().max_by(|a, b| {
-                a.success_rate
-                    .partial_cmp(&b.success_rate)
-                    .unwrap_or(std::cmp::Ordering::Equal)
-            }) {
-                if let Some(exp) = store.get_experiment(id).map_err(|e| {
-                    kernel_handle::KernelOpError::Internal(format!("Failed to get experiment: {e}"))
-                })? {
-                    if let Some(variant) = exp.variants.iter().find(|v| v.id == winner.variant_id) {
-                        let _ = store.set_active_version(variant.prompt_version_id, exp.agent_id);
-                        tracing::info!(
-                            experiment_id = %id,
-                            winner_variant = %winner.variant_name,
-                            success_rate = winner.success_rate,
-                            "Auto-activated winning variant's prompt version"
-                        );
-                    }
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    fn get_experiment_metrics(
-        &self,
-        experiment_id: &str,
-    ) -> Result<Vec<librefang_types::agent::ExperimentVariantMetrics>, kernel_handle::KernelOpError>
-    {
-        let id: uuid::Uuid = experiment_id.parse().map_err(|e| {
-            kernel_handle::KernelOpError::Internal(format!("Invalid experiment ID: {e}"))
-        })?;
-        let store = self
-            .prompt_store
-            .get()
-            .ok_or("Prompt store not initialized")?;
-        store.get_experiment_metrics(id).map_err(|e| {
-            kernel_handle::KernelOpError::Internal(format!("Failed to get experiment metrics: {e}"))
-        })
-    }
-
-    fn auto_track_prompt_version(
-        &self,
-        agent_id: librefang_types::agent::AgentId,
-        system_prompt: &str,
-    ) -> Result<(), kernel_handle::KernelOpError> {
-        let cfg = self.config.load();
-        if !cfg.prompt_intelligence.enabled {
-            return Ok(());
-        }
-        let store = self
-            .prompt_store
-            .get()
-            .ok_or(kernel_handle::KernelOpError::unavailable("Prompt store"))?;
-        match store.create_version_if_changed(agent_id, system_prompt, "auto") {
-            Ok(true) => {
-                tracing::debug!(agent_id = %agent_id, "Auto-tracked new prompt version");
-                // Prune old versions
-                let max = cfg.prompt_intelligence.max_versions_per_agent;
-                let _ = store.prune_old_versions(agent_id, max);
-                Ok(())
-            }
-            Ok(false) => Ok(()),
-            Err(e) => Err(kernel_handle::KernelOpError::Internal(format!(
-                "Failed to auto-track prompt version: {e}"
-            ))),
-        }
-    }
-}
-
-#[async_trait::async_trait]
-impl kernel_handle::WorkflowRunner for LibreFangKernel {
-    async fn run_workflow(
-        &self,
-        workflow_id: &str,
-        input: &str,
-    ) -> Result<(String, String), kernel_handle::KernelOpError> {
-        use crate::workflow::WorkflowId;
-        use kernel_handle::KernelOpError;
-
-        // Try parsing as UUID first, then fall back to name lookup.
-        let wf_id = if let Ok(uuid) = uuid::Uuid::parse_str(workflow_id) {
-            WorkflowId(uuid)
-        } else {
-            // Name-based lookup: scan all registered workflows.
-            let name_lower = workflow_id.to_lowercase();
-            let workflows = self.workflows.list_workflows().await;
-            workflows
-                .iter()
-                .find(|w| w.name.to_lowercase() == name_lower)
-                .map(|w| w.id)
-                .ok_or_else(|| {
-                    KernelOpError::Internal(format!("workflow `{}` not found", workflow_id))
-                })?
-        };
-
-        let (run_id, output) = LibreFangKernel::run_workflow(self, wf_id, input.to_string())
-            .await
-            .map_err(|e| KernelOpError::Internal(format!("Workflow execution failed: {e}")))?;
-
-        Ok((run_id.to_string(), output))
-    }
-}
-
-impl kernel_handle::GoalControl for LibreFangKernel {
-    fn goal_list_active(
-        &self,
-        agent_id_filter: Option<&str>,
-    ) -> Result<Vec<serde_json::Value>, kernel_handle::KernelOpError> {
-        let shared_id = shared_memory_agent_id();
-        let goals: Vec<serde_json::Value> =
-            match self.memory.structured_get(shared_id, "__librefang_goals") {
-                Ok(Some(serde_json::Value::Array(arr))) => arr,
-                Ok(_) => return Ok(Vec::new()),
-                Err(e) => {
-                    return Err(kernel_handle::KernelOpError::Internal(format!(
-                        "Failed to load goals: {e}"
-                    )))
-                }
-            };
-        let active: Vec<serde_json::Value> = goals
-            .into_iter()
-            .filter(|g| {
-                let status = g["status"].as_str().unwrap_or("");
-                let is_active = status == "pending" || status == "in_progress";
-                if !is_active {
-                    return false;
-                }
-                match agent_id_filter {
-                    Some(aid) => g["agent_id"].as_str() == Some(aid),
-                    None => true,
-                }
-            })
-            .collect();
-        Ok(active)
-    }
-
-    fn goal_update(
-        &self,
-        goal_id: &str,
-        status: Option<&str>,
-        progress: Option<u8>,
-    ) -> Result<serde_json::Value, kernel_handle::KernelOpError> {
-        let shared_id = shared_memory_agent_id();
-        let mut goals: Vec<serde_json::Value> =
-            match self.memory.structured_get(shared_id, "__librefang_goals") {
-                Ok(Some(serde_json::Value::Array(arr))) => arr,
-                Ok(_) => {
-                    return Err(kernel_handle::KernelOpError::Internal(format!(
-                        "goal `{goal_id}` not found"
-                    )))
-                }
-                Err(e) => {
-                    return Err(kernel_handle::KernelOpError::Internal(format!(
-                        "Failed to load goals: {e}"
-                    )))
-                }
-            };
-
-        let mut updated_goal = None;
-        for g in goals.iter_mut() {
-            if g["id"].as_str() == Some(goal_id) {
-                if let Some(s) = status {
-                    g["status"] = serde_json::Value::String(s.to_string());
-                }
-                if let Some(p) = progress {
-                    g["progress"] = serde_json::json!(p);
-                }
-                g["updated_at"] = serde_json::Value::String(chrono::Utc::now().to_rfc3339());
-                updated_goal = Some(g.clone());
-                break;
-            }
-        }
-
-        let result = updated_goal.ok_or_else(|| {
-            kernel_handle::KernelOpError::Internal(format!("goal `{goal_id}` not found"))
-        })?;
-
-        self.memory
-            .structured_set(
-                shared_id,
-                "__librefang_goals",
-                serde_json::Value::Array(goals),
-            )
-            .map_err(|e| {
-                kernel_handle::KernelOpError::Internal(format!("Failed to save goals: {e}"))
-            })?;
-
-        Ok(result)
-    }
-}
-
-impl kernel_handle::ToolPolicy for LibreFangKernel {
-    fn tool_timeout_secs(&self) -> u64 {
-        let cfg = self.config.load();
-        cfg.tool_timeout_secs
-    }
-
-    fn tool_timeout_secs_for(&self, tool_name: &str) -> u64 {
-        let cfg = self.config.load();
-        // 1. Exact match
-        if let Some(&t) = cfg.tool_timeouts.get(tool_name) {
-            return t;
-        }
-        // 2. Best glob match — longest pattern wins (most specific first).
-        // BTreeMap iteration is sorted by key; picking the longest matching
-        // pattern gives deterministic resolution when multiple globs match.
-        let best = cfg
-            .tool_timeouts
-            .iter()
-            .filter(|(pattern, _)| librefang_types::capability::glob_matches(pattern, tool_name))
-            .max_by_key(|(pattern, _)| pattern.len());
-        if let Some((_, &timeout)) = best {
-            return timeout;
-        }
-        // 3. Global fallback
-        cfg.tool_timeout_secs
-    }
-
-    fn skill_env_passthrough_policy(
-        &self,
-    ) -> Option<librefang_types::config::EnvPassthroughPolicy> {
-        let cfg = self.config.load();
-        librefang_types::config::EnvPassthroughPolicy::from_skills_config(&cfg.skills)
-    }
-
-    fn channel_file_download_dir(&self) -> Option<std::path::PathBuf> {
-        Some(self.config.load().channels.effective_file_download_dir())
-    }
-
-    fn effective_upload_dir(&self) -> std::path::PathBuf {
-        self.config_ref().channels.effective_file_download_dir()
-    }
-
-    fn readonly_workspace_prefixes(&self, agent_id: &str) -> Vec<std::path::PathBuf> {
-        self.named_workspace_prefixes(agent_id)
-            .into_iter()
-            .filter(|(_, mode)| *mode == WorkspaceMode::ReadOnly)
-            .map(|(p, _)| p)
-            .collect()
-    }
-
-    fn named_workspace_prefixes(&self, agent_id: &str) -> Vec<(std::path::PathBuf, WorkspaceMode)> {
-        let Ok(aid) = agent_id.parse::<AgentId>() else {
-            return vec![];
-        };
-        let Some(entry) = self.registry.get(aid) else {
-            return vec![];
-        };
-        if entry.manifest.workspaces.is_empty() {
-            return vec![];
-        }
-        let cfg = self.config.load();
-        let workspaces_root = cfg.effective_workspaces_dir();
-        let canonical_mount_roots =
-            workspace_setup::canonicalize_allowed_mount_roots(&cfg.allowed_mount_roots);
-        entry
-            .manifest
-            .workspaces
-            .iter()
-            .filter_map(|(name, decl)| {
-                workspace_setup::resolve_workspace_decl(
-                    name,
-                    decl,
-                    &workspaces_root,
-                    &canonical_mount_roots,
-                )
-            })
-            .collect()
-    }
-}
-
-// ---- BEGIN ApiAuth + SessionWriter impls (#3744) ----
-
-impl kernel_handle::ApiAuth for LibreFangKernel {
-    fn auth_snapshot(&self) -> kernel_handle::ApiAuthSnapshot {
-        // Single `config.load()` so every field in the snapshot comes from
-        // the same hot-reload generation. Per-request middleware would
-        // otherwise race the reload barrier and observe e.g. an old
-        // `api_key` alongside a new `dashboard_pass_hash`.
-        let cfg = self.config.load();
-        kernel_handle::ApiAuthSnapshot {
-            api_key: cfg.api_key.clone(),
-            dashboard: kernel_handle::DashboardRawConfig {
-                user: cfg.dashboard_user.clone(),
-                pass: cfg.dashboard_pass.clone(),
-                pass_hash: cfg.dashboard_pass_hash.clone(),
-            },
-            home_dir: self.home_dir().to_path_buf(),
-            device_api_keys: self.pairing.device_api_keys(),
-            config_users: cfg
-                .users
-                .iter()
-                .map(|u| kernel_handle::ApiUserConfigSnapshot {
-                    name: u.name.clone(),
-                    role: u.role.clone(),
-                    api_key_hash: u.api_key_hash.clone(),
-                })
-                .collect(),
-        }
-    }
-}
-
-impl kernel_handle::SessionWriter for LibreFangKernel {
-    fn inject_attachment_blocks(
-        &self,
-        agent_id: librefang_types::agent::AgentId,
-        blocks: Vec<librefang_types::message::ContentBlock>,
-    ) {
-        use librefang_types::message::{Message, MessageContent, Role};
-
-        let entry = match self.registry.get(agent_id) {
-            Some(e) => e,
-            None => {
-                tracing::warn!(
-                    agent_id = ?agent_id,
-                    "inject_attachment_blocks: agent not found in registry"
-                );
-                return;
-            }
-        };
-
-        let mut session = match self.memory.get_session(entry.session_id) {
-            Ok(Some(s)) => s,
-            _ => librefang_memory::session::Session {
-                id: entry.session_id,
-                agent_id,
-                messages: Vec::new(),
-                context_window_tokens: 0,
-                label: None,
-                messages_generation: 0,
-                last_repaired_generation: None,
-            },
-        };
-
-        let block_count = blocks.len();
-        let block_kinds: Vec<&'static str> = blocks
-            .iter()
-            .map(|b| match b {
-                librefang_types::message::ContentBlock::Image { .. } => "image",
-                librefang_types::message::ContentBlock::Text { .. } => "text",
-                librefang_types::message::ContentBlock::ImageFile { .. } => "image_file",
-                librefang_types::message::ContentBlock::ToolUse { .. } => "tool_use",
-                librefang_types::message::ContentBlock::ToolResult { .. } => "tool_result",
-                librefang_types::message::ContentBlock::Thinking { .. } => "thinking",
-                librefang_types::message::ContentBlock::Unknown => "unknown",
-            })
-            .collect();
-
-        session.push_message(Message {
-            role: Role::User,
-            content: MessageContent::Blocks(blocks),
-            pinned: false,
-            timestamp: Some(chrono::Utc::now()),
-        });
-
-        let total_messages_after = session.messages.len();
-
-        if let Err(e) = self.memory.save_session(&session) {
-            tracing::warn!(
-                agent_id = ?agent_id,
-                session_id = ?entry.session_id,
-                block_count,
-                error = %e,
-                "inject_attachment_blocks: failed to save session"
-            );
-        } else {
-            tracing::info!(
-                agent_id = ?agent_id,
-                session_id = ?entry.session_id,
-                block_count,
-                block_kinds = ?block_kinds,
-                total_messages_after,
-                "inject_attachment_blocks: injected content blocks into session"
-            );
-        }
-    }
-}
-
+//
+// All 16 `impl kernel_handle::* for LibreFangKernel` blocks now live in
+// `kernel::handles::*`. Each sub-module is a descendant of `kernel`, so
+// it retains access to `LibreFangKernel`'s private fields and inherent
+// methods without any visibility surgery. Specifically:
+//
+//   - `kernel::handles::agent_control`    — `kernel_handle::AgentControl`
+//   - `kernel::handles::memory_access`    — `kernel_handle::MemoryAccess`
+//   - `kernel::handles::task_queue`       — `kernel_handle::TaskQueue`
+//   - `kernel::handles::event_bus`        — `kernel_handle::EventBus`
+//   - `kernel::handles::knowledge_graph`  — `kernel_handle::KnowledgeGraph`
+//   - `kernel::handles::cron_control`     — `kernel_handle::CronControl`
+//   - `kernel::handles::hands_control`    — `kernel_handle::HandsControl`
+//   - `kernel::handles::approval_gate`    — `kernel_handle::ApprovalGate`
+//   - `kernel::handles::a2a_registry`     — `kernel_handle::A2ARegistry`
+//   - `kernel::handles::channel_sender`   — `kernel_handle::ChannelSender`
+//   - `kernel::handles::prompt_store`     — `kernel_handle::PromptStore`
+//   - `kernel::handles::workflow_runner`  — `kernel_handle::WorkflowRunner`
+//   - `kernel::handles::goal_control`     — `kernel_handle::GoalControl`
+//   - `kernel::handles::tool_policy`      — `kernel_handle::ToolPolicy`
+//   - `kernel::handles::api_auth`         — `kernel_handle::ApiAuth`
+//   - `kernel::handles::session_writer`   — `kernel_handle::SessionWriter`
+//
 // ---- END role-trait impls (#3746) ----
 
 // ---------------------------------------------------------------------------
